% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCircleOfCor.R
\name{addArrows}
\alias{addArrows}
\title{Add arrows to a factorial map
(best with a circle of Correlation).}
\usage{
addArrows(X, axis1 = 1, axis2 = 2, color = "darkorchid",
  alpha = 0.6, size = 1, center = c(0, 0), arrowLength = 0.3)
}
\arguments{
\item{X}{a data frame with the coordinates of the points
(e.g., from \code{ExPosition::epCA()}).}

\item{axis1}{(\code{default} = 1)
The number of the column of \code{X} for the horizontal axis.}

\item{axis2}{(\code{default} = 2)
The number of the column of \code{X} for the vertical  axis.}

\item{color}{the color of the arrows. Can be one color or a vector of colors.
If \code{color} is a vector, it needs to have exactly as many
elements as
the number of arrows  to plot.
\code{Default = 'darkorchid'}.}

\item{alpha}{(Default: 0.6),
the alpha (transparency) for the arrows, should be between 1
(no transparency) and 0 (completely transparent).}

\item{size}{(\code{Default = 1}), the thickness of the
segment}

\item{center}{(\code{Default: c(0, 0)}) the center of the graph}

\item{arrowLength}{(\code{Default: 0.3}), the lenth (in cm.) of
the terminal arrow.}
}
\value{
a \code{ggplot2} component to be added to a
scatterplot / map (typically created by
\code{createFactorMap()}).
}
\description{
\code{addArrows}: Add arrows to a factorial map
(i.e., a PCA-like map),
most likely  a plot of correlation produced by
\code{createFactorMap} \emph{et alia}.
}
\details{
The map should should first be created by, for example,
\code{createFactorMap()} (or equivalent functions from
\code{PTCA4CATA}), and then the arrows are added
(see example).  \code{addArrows} is typically used
with \code{\link{addCircleOfCor}}.
}
\examples{
\dontrun{
# Some PCA-like correlations
corXY <- matrix(c(.5,-.5, .1,.7, .8,.5, -.1,.9,  -.6,-.6),
               ncol = 2, byrow = TRUE )
# create a map of correlation
MapCor <- createFactorMap(corXY,
         constraints = list(minx = -1, miny = -1,
                            maxx = 1 , maxy = 1) )
# Add arrows and a circle to the base Map
ggMapWithCircle <- MapCor$zeMap +  addArrows(corXY) + addCircleOfCor()
# To print the map with arrows and circle:
# print(ggMapWithCircle)
}
}
\seealso{
\code{\link[ggplot2]{annotate}} \code{\link{addCircleOfCor}} \code{\link[PTCA4CATA]{createFactorMap}}
}
