% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiplot4PCA.R
\name{addArrowsAndNames}
\alias{addArrowsAndNames}
\title{Add arrows and names
to a factorial graph created by \code{createFactorMap}
and friends.}
\usage{
addArrowsAndNames(
  DATA,
  axis1 = 1,
  axis2 = 2,
  fontface = "bold.italic",
  color = "darkolivegreen"
)
}
\arguments{
\item{DATA}{a matrix or data frame with coordinates
of the items to be plotted
(NB \emph{need to have row names}).}

\item{axis1}{horizontal axis (\code{Default: 1}).}

\item{axis2}{vertical axis (\code{Default: 2}).}

\item{fontface}{font face for the names
(\code{Default: 'bold.italic'}).}

\item{color}{an element or
a vector of color (\code{Default: 'darkolivegreen'}).}
}
\value{
A list with two elements
\enumerate{
\item \code{arrows}  (\code{ggplot} code to add the arrows)
\item \code{labels4Biplot}
(\code{ggplot} code to add the names of the
items): Uses \code{ggplot2::annotate()}.
}
}
\description{
\code{addArrowsAndNames}:
Add arrows and names
to a factorial graph created by \code{createFactorMap}
and friends
}
\details{
More here later
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Hervé Abdi
}
