% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCircleOfCor.R
\name{addCircleOfCor}
\alias{addCircleOfCor}
\title{Add a Circle of Correlation to a PCA-like map
of correlation produced by \code{createFactorMap()} \emph{et alia.}}
\usage{
addCircleOfCor(color = "darkorchid", alpha = 0.3, size = 1,
  center = c(0, 0), radius = 1, nPoints = 100)
}
\arguments{
\item{color}{(\code{Default: 'darkorchid'}) color for the circle.}

\item{alpha}{(\code{Default: 0.3}) transparency for the circle,
should be between 0 (completely transparent) and 1
(no transparent).}

\item{size}{(\code{Default: 1}) thickness of the line of the circle.}

\item{center}{(\code{Default: c(0, 0)}) center of the circle.}

\item{radius}{(\code{Default: 1}) radius of the circle.}

\item{nPoints}{(\code{Default: 100}) the number of points used to
draw the circle.}
}
\value{
nothing
}
\description{
\code{addCircleOfCor}:
Add a Circle of Correlation to a PCA-like map
of correlation produced by \code{createFactorMap} \emph{et alia}.
}
\details{
The map should should be first created by, for example,
\code{createFactorMap()} (or equivalent functions from
\code{PTCA4CATA}), and then the circle of correlation is added
(see example).
}
\examples{
\dontrun{
# Some PCA-like correlations
corXY <- matrix(c(.5,-.5, .1,.7, .8,.5, -.1,.9,  -.6,-.6),
               ncol = 2, byrow = TRUE )
# create a map of correlation
MapCor <- PTCA4CATA::createFactorMap(corXY,
         constraints = list(minx = -1, miny = -1,
                            maxx = 1 , maxy = 1) )
# Add a circle to the base Map
ggMapWithCircle <- MapCor$zeMap  + addCircleOfCor()
# To print the map with the circle:
# print(ggMapWithCircle)
}
}
\seealso{
\code{\link[PTCA4CATA]{createFactorMap}}
}
\author{
Herve Abdi
}
