% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodies4MCA.R
\name{addLines4MCA}
\alias{addLines4MCA}
\title{(\code{ggplot2}) add lines to the levels of
the qualitative variables in a (variable) factorial
map of an MCA (e.g., computed with
\code{ExPosition::epMCA}).}
\usage{
addLines4MCA(
  Fj,
  col4Var,
  alpha = 0.7,
  linetype = 3,
  size = 0.5,
  axis_h = 1,
  axis_v = 2,
  dimension.names = "Dimension",
  ...
)
}
\arguments{
\item{Fj}{the output for the column set of an
MCA (e.g., from \code{ExPosition::epMCA}, this would be
\code{$ExPosition.Data$fj}).}

\item{col4Var}{vector of colors for the variables.
if col4Var is not equal to the number
of qualitative variables \code{ExPosition::epMCA})
will create these colors from
\code{\link[prettyGraphs]{prettyGraphsColorSelection}}.}

\item{alpha}{(Default: 0.7) the alpha values for the
lines.}

\item{linetype}{(Default: 3, dotted) the linetype number for the
lines.}

\item{size}{(Default: .5) the size of the lines.}

\item{axis_h}{(Default: 1) what is the horizontal axis.}

\item{axis_v}{(Default: 2) hat is the verical axis.}

\item{dimension.names}{(Default: \code{'Dimension'})}

\item{...}{Everything else that can be passed to
\code{ggplot2::geom_path()}.}
}
\value{
A list with the lines for each variables
}
\description{
\code{addLines4MCA}
is a \code{ggplot2} based function that
adds lines to join the levels of
the qualitative variables in a factorial
map of an MCA (e.g., computed with
\code{ExPosition::epMCA} and created with
\code{createFactortMap}).
}
\details{
The levels of a given variable
are first ordered by alphabetical order prior
to drawing the lines
(so, e.g., \code{truc.1}, \code{truc.2}
and \code{truc.3} will be correctly ordered).

The name (i.e., stem) of a variable is obtained
by stripping the names of the levels (i.e., columns)
of the extension after the last "." of their names.
For example, the name \code{toto.1} for
a row of \code{Fj} corresponds to the qualitative
variable \code{toto}.
}
\section{Important_Note}{


When creating multiple layers graphs,
 because of the way \code{ggplot2} creates graphs,
  all the matrices/dataframe should all
  the have the same column names
  [e.g., \code{colnames()}
  equal to c("Dimension 1", "Dimension 2")].
   When this is not the case,
   some strange and cryptic error may be produced
(e.g., "cannot find Dimension").
}

\examples{
\dontrun{
if(interactive()){
library{ExPosition}
data("mca.wine")
resMCA    <- epMCA(mca.wine$data, graphs = FALSE)
Fj <- resMCA$ExPosition.Data$fj
baseMap4J <- createFactorMap(Fj)$zeMap +
  addLines4MCA(Fj,col4Var = rep('red',10) )
 }
}
}
\seealso{
\code{\link[prettyGraphs]{prettyGraphsColorSelection}}
 \code{\link{getVarNames}}
}
