% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcoholInEurope.R
\docType{data}
\name{alcoholInEurope}
\alias{alcoholInEurope}
\title{The average consumption of 4 types of alcoholic beverages
in 22 European countries and 3 neighboring countries.}
\format{A list containing 2 data frames:
\itemize{
\item \code{df.active:} the 22 countries by
4 types of alcoholic beverages active data set.
\item \code{supplementary.observations:}
the 3 (non-EU but neighoring) countries by
4 types of alcoholic beverages supplementary data set.
}}
\usage{
data("alcoholInEurope")
}
\description{
These data
give the amount in liters per year and per person older than 15
for 4 categories of alcoholic beverages:
(\code{Beer, Wine, Spirit, Other}) in 22 European countries and 3
neighboring countries.
These data were obtained from an article originally published
in the journal
\emph{le Monde} dated September 25, 2018.
}
\references{
These data were obtained from the \emph{World Health Organization}
and were collected and presented in an article from the journal
\emph{le Monde} dated September 25, 2018. See the following link
for this article and details:

\url{https://www.lemonde.fr/les-decodeurs/article/2018/08/24/quels-alcools-preferent-boire-les-europeens_5345933_4355770.html}
}
\author{
Hervé Abdi from Le Monde
}
\keyword{data4PCCAR}
\keyword{datasets}
