% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindSortingWines.R
\docType{data}
\name{blindSortingWines}
\alias{blindSortingWines}
\title{Wine experts and panelists
(blind) sorted red, rosé, and white wines.
These data can be used to illustrate
multidimensional scaling methods
(e.g., MDS or DISTATIS).}
\format{
A list containing
five data frames
\itemize{
\item \code{wineInformation: } a df
describing the 18 wines. The colors
are coded \code{P, R, W}
(Rosé, Red, and White).
\item \code{distance.Experts:} the 18 by 18
distance matrix derived from the
sorting of the 26 wine Experts.
\item \code{distance.Panelists:} the 18 by  18
distance matrix derived from the
sorting of the 26 wine Panelists.
\item \code{df.Experts:} The
18 wines (rows) by
26 wine Experts (columns)
data frame storing the results of the
sorting task by the wine
Experts (from \code{S1} to \code{S26}).
Here two wines (rows)
with the same number were
placed in the same group by the
Expert
(columns).
\item \code{df.Experts:} The
18 wines (rows) by 19
wine Panelists (columns)
data frame storing the results of the
sorting task by the wine
Panelists (from \code{J1} to \code{J19}).
Here two wines (rows)
with the same number were
placed in the same group by the
Panelist
(columns).
}
}
\usage{
data("blindSortingWines")
}
\description{
\code{blindSortingWines}
26 wine Experts and 19 Panelists (i.e., novices)
blind sorted 18 wines
(6 red, 6 rosé, and 6 white).

The 26 Experts were told to sort the
wines into three groups whereas the
Panelists just told to make as many groups
as they wished but less the 18 and more than 1.
}
\references{
These data (and the story around)
are described in more details in
Ballester, J., Abdi, H.,
Langlois, J., Peyron, D.,
& Valentin, D. (2009).
The odor of colors:
Can wine experts and novices
distinguish the odors of
white, red, and rosé wines?
\emph{Chemosensory Perception}, \strong{2}, 203-213.
}
\author{
Jordi Ballester, Dominique Valentin, &
Hervé Abdi
}
\keyword{data4PCCAR}
\keyword{datasets}
