% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMC.R
\name{boot.eigen}
\alias{boot.eigen}
\title{generate a bootstrap distribution of eigenvalues
for a given matrix}
\usage{
boot.eigen(X, nIter = 100, center = TRUE, scale = "SS1")
}
\arguments{
\item{X}{The original data matrix.}

\item{nIter}{how many bootstrapped sets of eigenvalues
to generate; Default: \code{100}.}

\item{center}{(Default = \code{TRUE}) if
\code{TRUE}: center the data (by column).}

\item{scale}{the type of scaling of the columns of
data. Can be \code{FALSE} (no scaling),
\code{TRUE} (scale as \eqn{Z}-scores
with squared norm = \eqn{I - }1),
 or \code{'SS1'} (all columns
of the data matrix have norm 1, and so the eigenvalues
come from a correlation matrix).
Default: \code{'SS1'}.}
}
\value{
a list with 3 elements
\enumerate{
\item \code{$fixed.eigs}: the eigen-values of \eqn{X},
\item \code{$boot.eigs}: an \code{nIter} by rank(\eqn{X}) matrix
of the eigenvalues of the bootstrapped samples,
\item \code{$boot.eigs.sorted}: an \code{nIter} by rank(\eqn{X}) matrix
of the eigenvalues of the bootstrapped samples.
}
}
\description{
\code{boot.eigen}:
generates
 a bootstrap distribution of eigenvalues
for a given (rectangular) matrix.
}
\details{
\code{boot.eigen} uses \code{scale0} to
normalize the data and so will not create \code{NaN}
when the variance of a column is 0 (it will create
a vector of 0s).
}
\examples{
data(iris)
bootstapped.Eigenvalues <- boot.eigen(iris[,1:4], nIter = 10)
}
\seealso{
\code{\link{scale0}} \code{\link{monteCarlo.eigen}}
}
\author{
Hervé Abdi
}
