% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot4CCA.R
\name{Boot4CCA}
\alias{Boot4CCA}
\title{Boot4CCA}
\usage{
Boot4CCA(DATA1, DATA2, center1 = TRUE, center2 = TRUE,
  scale1 = "ss1", scale2 = "ss1", Fi = NULL, Fj = NULL,
  nf2keep = 3, nIter = 1000, critical.value = 2, eig = FALSE,
  alphaLevel = 0.05)
}
\arguments{
\item{DATA1}{}

\item{DATA2}{}

\item{center1}{}

\item{center2}{}

\item{scale1}{}

\item{scale2}{}

\item{Fi}{}

\item{Fj}{}

\item{nf2keep}{}

\item{nIter}{}

\item{critical.value}{}

\item{eig}{}

\item{alphaLevel}{}
}
\description{
Boot4CCA
}
\examples{
\dontrun{}
}
