% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferences4PLSC.R
\name{Boot4PLSC}
\alias{Boot4PLSC}
\title{Create a Bootstrap Cube for PLSC}
\usage{
Boot4PLSC(DATA1, DATA2, center1 = TRUE, center2 = TRUE,
  scale1 = "ss1", scale2 = "ss1", Fi = NULL, Fj = NULL,
  nf2keep = 3, nIter = 1000, critical.value = 2, eig = FALSE,
  alphaLevel = 0.05)
}
\arguments{
\item{DATA1}{an \eqn{N*I}  data matrix}

\item{DATA2}{an \eqn{N*J}  data matrix
(measured on the same observations as \code{DATA2})}

\item{center1}{when \code{TRUE} (default) \code{DATA1}
will be centered}

\item{center2}{when \code{TRUE} (default) \code{DATA2}
will be centered}

\item{scale1}{when \code{TRUE} (default) \code{DATA1}
will be normalized. Depends upon \code{ExPosition}
function \code{expo.scale} whose description is:
boolean, text, or (numeric) vector.
If boolean or vector,
it works just as scale.
The following text options are available:
\code{'z'}: \eqn{z}-score normalization,
\code{'sd'}: standard deviation normalization,
\code{'rms'}: root mean square normalization,
 \code{'ss1'}: sum of squares
 (of columns) equals 1
 (i.e., each column vector has length of 1).}

\item{scale2}{when \code{TRUE} (default) \code{DATA2}
will be normalized
 (same options as for \code{scale1}).}

\item{Fi}{= \code{NULL}, the \eqn{I} factor scores
for the columns of \code{DATA1}.
if \code{NULL}, the function computes them.}

\item{Fj}{= \code{NULL}, the J factor scores
for the columns of \code{DATA2}.
if \code{NULL} the function computes them.
\code{Boot4RowCA} will compute them.}

\item{nf2keep}{How many factors to
keep for the analysis (Default = \code{3}).}

\item{nIter}{(Default = \code{1000}). Number of Iterations
(i.e., number of Bootstrtap samples).}

\item{critical.value}{(Default = \code{2}).
The critical value for a \code{BR} to be considered
significant.}

\item{eig}{if \code{TRUE} compute bootstraped
confidence intervals (CIs) for the eigenvalues
(default is \code{FALSE}).}

\item{alphaLevel}{the alpha level used to compute
the confidence intervals for the eigenvalues
(with CIS at 1-alpha). Default is \code{.05}}
}
\value{
a list with
\itemize{
\item{\code{bootstrapBrick.i}}{the
     Bootstraped factor scores  for the \eqn{I}-set;}
 \item{\code{bootRatios.i}}{the bootstrap ratios
     for the \eqn{I}-set;}
\item{\code{bootRatiosSignificant.i}}{the Significant
   BRs for the \eqn{I}-set;}
\item{\code{bootCube.j}}{
    An \code{Items * Dimensions * Iterations} Brick of
    Bootstraped factor scores for the \eqn{J}-set;}
\item{\code{bootRatios.j}}{the bootstrap ratios for the \eqn{J}-set;}
\item{\code{bootRatiosSignificant.j}}{the Significant
   BRs for the \eqn{J}-set;}
   }
   In addition if \code{eig = TRUE}, the list includes:
\itemize{
\item{\code{eigenValues}}{the \code{nIter * nL} table
 of eigenvalues;}
 \item{\code{eigenCIs}}{the CIs for the
 eigenvalues.}
 }
}
\description{
\code{Boot4PLSC}:
Creates Bootstrap Cubes for the \eqn{I} and \eqn{J} sets
of a PLSC. The cubes are
obtained from bootstraping the rows
of the two data-tables used for PLSC.
Uses the "transition formula" to get
the values of the rows and columns loadings
from multiplication of the latent variables.
Gives also the bootstraped eigenvalues
(if \code{eigen = TRUE}).
}
\details{
Planned development: A compact version that gives only
bootstrap ratios (not BootstrapBricks).
}
\author{
Hervé Abdi
}
