% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomImage4PCA.R
\name{buildRandomImage4PCA}
\alias{buildRandomImage4PCA}
\title{Make a random image (with similar PCA loadings) of a
matrix of integers data.}
\usage{
buildRandomImage4PCA(X, nObs = NROW(X), center = TRUE, scale = FALSE)
}
\arguments{
\item{X}{The original data}

\item{nObs}{how many observations for the random data
(Default: \code{NROW(X)}.}

\item{center}{if \code{TRUE} (Default) perform a
centered PCA of \code{X}.}

\item{scale}{if \code{FALSE}
(Default)  do not scale the data for the PCA of \code{X}.
Other options are \code{TRUE} (use \eqn{Z}-scores), \code{'SS1'},
normalize each variable to length 1.}
}
\value{
A matrix of (somewhat) random integers with
\code{nObs} observations whose
factorial structure roughly matches the original matrix.
}
\description{
\code{buildRandomImage4PCA}:
Makes a random image (with similar PCA loadings) of a
matrix of integer data such as Likert scales data..
}
\details{
The procedure starts with a PCA of the original matrix
\code{X} and gets loadings and factor scores.
From the eigenvalues,
 \code{buildRandomImage4PCA} generates random observations in
 the factor scores spaces. These observations are then used
 to recreate (via the \emph{reconstitution formula})
 a new matrix \code{X.rand} that will store the observations
 that will recreatae the random factor scores. The
 observation in \code{X.rand} are then re-scaled
 to match the distribution of the integers in the original matrix.
 Doing so gives a matrix of scores with the same basic statistics
 (i.e., mean and standard deviation) per column whose PCA will give
 roughly the same loadings (not the same factor score of course)
 as the original matrix.
}
\examples{
library(ExPosition)
data("twentyWines")
test4build <- buildRandomImage4PCA(twentyWines$df.active)
}
\seealso{
\code{\link[ExPosition]{epPCA}}
}
\author{
Hervé Abdi
}
