% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodies4MCA.R
\name{coloringLevels}
\alias{coloringLevels}
\title{create a vector of color names for the levels
of an MCA from the color names of the variables.}
\usage{
coloringLevels(levelsNames, colorOfVariables = NULL)
}
\arguments{
\item{levelsNames}{The names of the levels of the
columns of an MCA as run, for example, by
\code{ExPosition::epcMCA}.
Could be, for example the names of the \code{fj}
factor scores.}

\item{colorOfVariables}{A vector of colors
for the variables (should match the values obtained from the
levels).
if \code{NULL} (default)
or if the numbers of variables do not match the
number derived from \code{levelsNames} then
get the colors from
\code{}}
}
\value{
A list with 2 vectors:
\code{color4Levels} the colors of the labels
and \code{color4Variables} the colors of the
variables (in the order in which they appeared).
}
\description{
\code{coloringLabels}:
creates a vector of color names for the levels
(i.e., columns)
of an MCA from the color names of the variables.
}
\details{
If \code{colorOfVariables} is provided
the colors are used in the order in which the  variables
appear in the list of the levels.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[prettyGraphs]{prettyGraphsColorSelection}}
 \code{\link{getVarNames}}
}
\author{
Hervé Abdi
}
