% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOfMusic.R
\docType{data}
\name{colorOfMusic}
\alias{colorOfMusic}
\title{12 Children and 10 Adults picked up
the color that best describes 9 pieces of music.}
\format{A list with 5 data frames or arrays
describing respectively
1) the check mark data, 2) the description of the participants,
3) the names and color codes of the 10 colors used,
4) the binary cube of data  (color by music by participants),
and 5)
the (pseudo) contingency table of data (color by music).
\describe{
\item{\code{participantsChoice}}{A 22 rows (Participants) by 9
columns (Pieces of Music) data frame. The number
at the intersection of
of a row and a column gives the number-code
(from 1 to 10) of the chosen color
for the music (column) by the participant (row).
The name of the colors are given in the data frame stored in
\code{$colorInformation}.}
\item{\code{participantsDescription}}{A 22 by 2
data frame  describing the
participants according to \code{Age} (Child vs Adult) and
\code{Gender} (F vs M).}
\item{\code{colorInformation}}{The name
of the colors and their color code (useful when plotting the
data).}
\item{\code{cubeOfData}}{The 10-Colors by 9-Music pieces
 by 22-Participants
 cube of 0/1 data. A value of 1 (resp. 0) means that the participant
 chose (resp. did not choose)
 the color to match the piece of music}
\item{\code{contingencyTable}}{The 10-Colors
 by 9-Music Pieces (pseudo) contingency table.
 The value at the intersection of a row (color) and a column
 (piece of music) is the number of participants who chose this
 color to match this piece of music. This contingency
 table is typically analyzed with correspondence analysis.
 (e.g., see Abdi and Bera, 2018). }
}}
\source{
Abdi, H. and Bera, M. (2018).  Correspondence Analysis.
\emph{Encyclopedia of Social Network Analysis and Mining} (2nd Edition).
New York: Springer Verlag.
\url{www.utdallas.edu/~herve}.
}
\usage{
data("colorOfMusic")
}
\description{
\code{colorOfMusic}. A data set:
 12 Children and 10 Adults picked up
one color to describe 9 pieces of music (one at a time).
The participants are described by their
age (children vs adults) and by their gender
(F vs M). This data set is used to illustrate
correspondence analysis in Abdi and Bera (2018).
}
\author{
Herve Abdi
}
\keyword{data4PCCAR}
\keyword{datasets}
