% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorsOfVowels.R
\docType{data}
\name{colorsOfVowels}
\alias{colorsOfVowels}
\title{133 (French) participants
 associated 6 colors to 6 vowels (data from
Chastaing, 1961).}
\format{
A list containing 5 objects:
1) \code{CT}: a 6 vowels by 6 colors (pseudo) contingency table
storing the responses of all 133 participants;
2) \code{CT.noSynesthesia}: a 6 vowels by 6 colors (pseudo)
contingency table
storing the responses of the 27 non synesthesia participants;
3) \code{CT.Synesthesia}: a 6 vowels by 6 colors (pseudo)
contingency table
storing the responses of the 106 synesthesia participants;
4) \code{vowelsDescription} a data frame describing the
vowels (from their place of articulation as anterior or posterior);
and
5) \code{colorsDescription} a data frame
giving the R-color code for the colors.
 @keywords datasets data4PCCAR
 @author Maxime Chastaing and Hervé Abdi
 @references
 The data reported here were obtained from Tables 1 and 2 from
  Chastaing, M (1961). des sons et des couleurs.
  \emph{Vie et Language, 112}, 358-365.
}
\usage{
data("colorsOfVowels")
}
\description{
\code{colorsOfVowels}
133 French
participants (106 with synesthesia and 27 without synesthesia)
associated 6 colors to 6 vowels (data from
Chastaing, 1961).
}
\details{
The participants were asked if they
had previously spontaneously associated a color to a vowel.
The 106 participants who responded "Yes" are called
with \emph{synesthesia} and the 27
participants who responded "No" are called
with \emph{without synesthesia}.

The question asked to the participants was:
``Here are six colors: red, orange, yellow, blue, and violet.
What vowel do you think match each color?".
So, the color are the stimuli
(i.e., the \emph{independent} variable)
and the vowels are the responses
(i.e., the \emph{dependent} variable).

The data were obtained by concatenation of the data from
Tables 1 and 2 of Chastaing (1961).
The vowels correspond to the following sound(s)
1) \emph{i} as in prett\strong{y},
2) \emph{y} the French "u"-sound as in t\strong{u}t\strong{u},
3) \emph{e} this column concatenate all the
accented e-sounds in French such as é, è and ê
(e.g., as in m\strong{ay}),
4) \emph{a} like P\strong{a}p\strong{a},
5) \emph{o} the o and ô sounds, and
6) \emph{u} the "ou" sound (like in y\strong{ou}).
}
