% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS_jack_svds_HA.R
\name{corrcoef4mat}
\alias{corrcoef4mat}
\title{FUNCTION_TITLE}
\usage{
corrcoef4mat(Y1, Y2)
}
\arguments{
\item{Y1}{PARAM_DESCRIPTION}

\item{Y2}{PARAM_DESCRIPTION}
}
\value{
A list with \eqn{R}2 and the \eqn{RV} coefficient between
\code{Y1} and \code{Y2}.
}
\description{
\code{corrcoef4mat}
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
