% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodies4MCA.R
\name{ctr4Variables}
\alias{ctr4Variables}
\title{Compute contributions (or squared cosines)
for (qualitatitve)  variables
in Multiple Correspondence Analysis (e.g., as performed by
\code{ExPosition::epMCA}).}
\usage{
ctr4Variables(ctrJ)
}
\arguments{
\item{ctrJ}{a matrix or data frame of contributions
or squared cosines
(e.g., from \code{ExPosition::epMCA}, the contributions
are in \code{cj}, the squared cosines are in \code{ri}).}
}
\value{
A qualitative variables by dimensions data frame.
}
\description{
\code{ctr4Variables}:
Computes contributions
(or squared cosines)  for (qualitatitve)  variables
in Multiple Correspondence Analysis (e.g., as performed by
\code{ExPosition::epMCA}).
}
\details{
In MCA, the contribution
(resp. squared cosine) of a variable is
the sum of (resp. squared cosine) contributions
of all its levels.

\code{ctr4Variables} finds the levels of a given variable
by stripping the contribution columns names of their extension
(e.g., \code{toto.1} and \code{toto.2} are two levels
of the qualitative variable \code{toto}). This
is performed with the function
\code{tools::file_path_sans_ext}.
}
\examples{
library(ExPosition)
data(mca.wine)
resMCA    <- epMCA(mca.wine$data, graphs = FALSE)
contriVar <- ctr4Variables(resMCA$ExPosition.Data$cj)
}
\seealso{
\code{\link{getVarNames}}
}
\author{
Hervé Abdi
}
