% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4PCCAR.R
\docType{package}
\name{data4PCCAR}
\alias{data4PCCAR}
\alias{data4PCCAR-package}
\title{\code{data4PCCAR:} Some  companion data sets for
Abdi, H. and Beaton, D. (2018)
\emph{Principal Component and Correspondence Analyses with R}.
New York: Springer.}
\description{
\code{data4PCCAR:} Some data sets to illustrate
and complement:  Abdi and Beaton(2018)
\emph{Principal Component and Correspondence Analyses with R}.
New York: Springer.
}
\details{
These data sets are used to illustrate principal component analysis,
correspondence analysis, and multidimensional scaling.

\tabular{ll}{ Package: \tab datarPCCAR\cr Type: \tab Package\cr Version: \tab
0.1.0\cr Date: \tab 2018-03-15\cr License: \tab GPL-2\cr Depends: \tab
 (>= 3.0.0)}
}
\author{
Herve Abdi [aut, com, ctb],
Vincent Guillemot  [aut, cre],
& Derek Beaton
[aut, cre]

Maintainer: Herve Abdi <herve@utdallas.edu>
}
\keyword{package}
