% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadInTheUSA.R
\docType{data}
\name{deadInTheUSA}
\alias{deadInTheUSA}
\title{The causes of death as a function of age in the USA in 2001.}
\format{A list containing 1 data frame:
\itemize{
\item \code{CT:} the contingency table
giving the number of death per age and cause.
}}
\usage{
data("deadInTheUSA")
}
\description{
\code{deadInTheUSA}: provides a
19 causes of death by 11 age-bins contingency table
storing the number of persons who died for each
cause for each age in the USA in 2001.
This type of data is typically
analyzed with correspondence analysis.
}
\details{
The data have been cleaned and compacted
from the original data to make the data table more readable.
The cause \code{AD_PD} is the concatenation of
Alzheimer and Parkinson diseases.
}
\references{
Original data from
\emph{National Vital Stat Reports, 52(3)}, Sept 18.
}
\author{
Hervé Abdi
}
\keyword{data4PCCAR}
\keyword{datasets}
