% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastInferences4CA.R
\name{eig4CA}
\alias{eig4CA}
\title{A bare-bones function to compute  the
eigen-values (and if asked row and column factor scores)
of a matrix suitable for correspondence analysis.}
\usage{
eig4CA(Xdata, eig.only = TRUE)
}
\arguments{
\item{Xdata}{a data matrix
(whose all entries are non-negative) suitable
for correspondence analysis.}

\item{eig.only}{when \code{TRUE} (Default)
compute only the CA- eigen-values of \code{X}.
Otherwise compute also the row and column
CA factor scores (i.e., \code{fi} and \code{fj}).}
}
\value{
if \code{eig.only} is \code{TRUE}
\code{eig4CA} returns the CA-eigenvalues of
\code{X};
if \code{eig.only} is \code{FALSE}
\code{eig4CA} returns
a list with: \code{$eigen}: the CA-eigenvalues of
\code{X}, \code{$fi}: the CA row factor scores,
and \code{$fi}: the CA column factor scores.
}
\description{
\code{eig4CA}:
A very fast and bare-bones function that computes
the eigenvalues
(and possibly the row and column factor scores)
of the Correspondence Analysis (CA)
of a data matrix suitable for CA
(i.e., a matrix whose all entries are non-negative).
\code{eig4CA} is mainly used for cross-validation
and resampling methods (e.g., permutation and
bootstrap tests).
}
\details{
As a fast bare-bones CA based computations
\code{eig4CA} is mainly used for
cross-validation for CA methods.
}
\examples{
\dontrun{
set.seed(87) # set the seed
X <- matrix(round(runif(21)*20), ncol = 3) # good for CA
eigenOfX <- eig4CA(X)
}
}
\author{
Hervé Abdi
}
