% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastInferences4-eigen-CA.R
\name{eigCA4Multinom}
\alias{eigCA4Multinom}
\title{Sample from a multinomial distribution
(with a given probability distribution)
and compute the eigenvalues
of a correpondence analysis of the simulated
matrix..}
\usage{
eigCA4Multinom(nobs, prob, nrow, ncol)
}
\arguments{
\item{nobs}{grand total of the table to be simulated.}

\item{prob}{probability distribution
for the cells. Should be length = \code{nrow*ncol}
(see below for these parameters).}

\item{nrow}{The number of rows of the matrix
to be simulated.}

\item{ncol}{The number of columns of the matrix
to be simulated.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
\code{eigCA4Multinom}:
Sample from a multinomial distribution
(with a given probability distribution)
and compute the eigenvalues
of the CA of an \code{nrow*ncol}
(see below for these parameters).
data matrix simulating correspondence analysis.
}
\details{
\code{eigCA4Multinom}
is mostly used for computing eigenvalues
of
created data matrices
simulating permutation and bootstrap procedures
for correspondence analysis.
}
\examples{
\dontrun{
set.seed(87) # set the seed
X <- matrix(round(runif(21)*20), ncol = 3) # good for CA
nobs <-  sum(X) # grand total
nI <- nrow(X)
nJ <- ncol(X)
pI <- as.matrix(rowSums(X) / nobs) # marginal I & J
pJ <- as.matrix(colSums(X) / nobs) # probabilites
p4Permutation <- pI \%*\% t(pJ) # Independence <=> permutation
# Simulated Permutation Probabilities
permEigen <- eigCA4Multinom(nobs, p4Permutation, nI, nJ)
p4Bootstrap   <- X / nobs # Actual prob <=> Bootstrap
permBoots <- eigCA4Multinom(nobs, p4Bootstrap, nI, nJ)
}
}
\author{
Hervé Abdi
}
