% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_internal_functions.R
\name{epGPCA2}
\alias{epGPCA2}
\title{epGPCA2, authored by Derek Beaton. 
Cf. package GSVD on GitHub, soon available on the CRAN}
\usage{
epGPCA2(DATA, scale = TRUE, center = TRUE, DESIGN = NULL,
  make_design_nominal = TRUE, masses = NULL, weights = NULL,
  graphs = TRUE, k = 0)
}
\description{
epGPCA2, authored by Derek Beaton. 
Cf. package GSVD on GitHub, soon available on the CRAN
}
\keyword{internal}
