% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_internal_functions.R
\name{epOutputHandler}
\alias{epOutputHandler}
\title{epOutputHandler, authored by Derek Beaton. 
Cf. package GSVD on GitHub, soon available on the CRAN}
\usage{
epOutputHandler(res = NULL, epPlotInfo = NULL)
}
\description{
epOutputHandler, authored by Derek Beaton. 
Cf. package GSVD on GitHub, soon available on the CRAN
}
\keyword{internal}
