% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epVari.R
\name{epVari}
\alias{epVari}
\title{Varimax rotation on loadings and factor scores.}
\usage{
epVari(resExpo, dim2keep = 2, normalize = TRUE, ...)
}
\arguments{
\item{resExpo}{the results of a principal component
analysis performed
by  \code{ExPosition::epPCA}.}

\item{dim2keep}{\code{(Default = 2)}
the number of dimensions to keep.}

\item{normalize}{(\code{default = TRUE}).
Passed to \code{\link[stats]{varimax}}:
logical. Should Kaiser normalization be performed?
If so the loadings are re-scaled
to unit length before rotation,
and scaled back afterwards.}

\item{...}{stuff to pass to \code{varimax}.}
}
\value{
A list with 6 elements:
\enumerate{
\item \code{rotationMatrix:} the rotation matrix.
\item \code{scaled.rotationMatrix:}
the scaled rotation matrix
(differs from \code{rotationMatrix}
only when \code{normalize = TRUE}), see
\code{\link[stats]{varimax}}
for more.
\item \code{rotated.I:} the rotated factor scores
(\eqn{I}-set).
\item \code{rotated.J:} the loadings (\eqn{J}-set).
\item \code{rotated.eigs:} the rotated pseudo-eigenvalues.
\item \code{rotated.t:}
the rotated percentage of explained inertia.
}
}
\description{
\code{epVari}: a
Varimax rotation on loadings and factor scores.
\code{epVari} is based on \code{stats::varimax}
(with parameters \code{normalize = TRUE, eps = 1e-6}), but
gives also the rotated factor scores, pseudo eigenvalues
and explained percentage
of inertia for the rotated dimensions.
}
\details{
\code{epVar} is a wrapper
around \code{stats::varimax}.
}
\examples{
data(iris)
library(ExPosition)
resPCA    <- ExPosition::epPCA(iris[,1:4], graphs = FALSE)
iris.Vari <-  epVari(resPCA)
}
\seealso{
\code{\link[stats]{varimax}}
}
\author{
Hervé Abdi
}
