% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferencesMultinom4CA-3.R
\name{fastPerm4CA}
\alias{fastPerm4CA}
\title{Permutation test for CA
on real contingency tables
obtained from a multinomial distribution.}
\usage{
fastPerm4CA(X, nIter = 1000, compact = FALSE)
}
\arguments{
\item{X}{the data matrix (non-negative numbers)}

\item{nIter}{(Default = 1000). Number of Iterations
(i.e. number of permuted samples computed).}

\item{compact}{if TRUE return only
\emph{p}-values for omnibus test
default is \code{FALSE}.}
}
\value{
a list with
\code{fixedInertia}: the CA-inertia of the data matrix;
\code{fixedEigenvalues}: the CA-eigenvalues of
the data matrix;
\code{pOmnibus}: the probability associated
to the inertia.
If \code{compact} is \code{FALSE}, return also
\code{permInertia}:
an \code{nIter} * 1 vector containing the
permuted inertia;
\code{pEigenvalues}: The probabilities
associated to each eigenvalue;
If \code{compact} is is \code{FALSE}, return also
\code{permEigenvalues}: an
\code{nIter} * \code{L} matrix giving
the permuted eigenvalues.
}
\description{
\code{fastPerm4CA} computes a permutation test
for CA when CA is performed
on a real contingency table.
The resampling is obtained
from a multinomial distribution.
\code{fastPerm4CA}
can be used for big tables
for the omnibus test (i.e., inertia) and for
the eigenvalues.
}
\author{
Hervé Abdi
}
