% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpos.R
\name{firstpos}
\alias{firstpos}
\title{firstpos}
\usage{
firstpos(P, Q = NULL)
}
\arguments{
\item{P}{A matrix of numeric values}

\item{Q}{If not NULL, a matrix of numeric values with the same number of columns as \code{P}}
}
\value{
If Q is NULL, then the function returns a only the modified P matrix.
Otherwise, the functions returns a list with two elements:
\itemize{
  \item \code{P} A matrix with the same values as in P, except that the sign of all the values in each column is flipped if the first value of this column is negative.
  \item \code{Q} A matrix with the same values as in Q, except that the sign of all the values in each column is flipped the same way that it is flipped for P.
}
}
\description{
Changes the sign of the values in two matrices according to the sign of the elements in the first line of the first matrix. This is used to get reproducible results when performing singular value decompositions or eigenvalue decompositions.
}
\examples{
\dontrun{
if(interactive()){
 A <- matrix(rnorm(4), 2, 2)
 B <- matrix(rnorm(6), 3, 2)
 firstpos(A, B)
 }
}
}
