% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpos4Expo.R
\name{firstpos4ExPo}
\alias{firstpos4ExPo}
\title{Makes positive the first value of factor scores and
singular vectors from the output of functions (e.g.,
\code{epPCA, epCA}) from
the \code{ExPosition} package.}
\usage{
firstpos4ExPo(resFile)
}
\arguments{
\item{resFile}{The result of an analysis of a function
from the \code{ExPosition} package (e.g., \code{epPCA}, \code{epCA})}
}
\value{
An \code{ExPosition} object with the
corrected results.
}
\description{
\code{firstpos4ExPo}: makes positive the first element
of the column factor scores (i.e., \code{fj})
and column singular vectors (i.e., \code{pdq$q}) of  analyses
performed with the  \code{ExPosition} package.
\code{firstpos4ExPo} also insures that the sign of the column
factor scores and singular vectors agree with
the row factors scores (i.e., \code{fi})
and singular vectors (i.e., \code{pdq$p}).

The transformation implemented by \code{firstpos4ExPo}
makes it easier to compare results from different analyses
because the sign of the factor scores
and singular vectors is arbitrary (i.e., a singular vector
multiplied by -1 is still a singular vector).
}
\details{
Note that only the results of the
computation are affected by the \code{firstpos4ExPo},
the graphics are \emph{not} affected. So when using
\code{firstpos4ExPo} it is wise to set the option
\code{graphs = FALSE}.
Also, a similar
version of \code{firstpos4ExPo} for the \code{InPosition}
family is not yet available.
}
\examples{
\dontrun{
data("mtcars") # load the mtcars data set 
resPCA.mtcars <- firstpos4ExPo(epPCA(mtcars, graphs = FALSE, k = 3))
# Check that the first element of the column vectors is positive
resPCA.mtcars$resPCA.mtcars$ExPosition.Data$fj
}
}
\seealso{
\code{\link{firstpos}},  \code{\link{renormInertiaExPo}}
}
\author{
Hervé Abdi
}
