% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveBeers5Judges.R
\docType{data}
\name{fiveBeers5Judges}
\alias{fiveBeers5Judges}
\title{Five tasters evaluated
five beers on dimensions
that they  had previously chosen.
These data can be used to illustrate,
for example,
Multiple Factor Analysis (MFA) or Distatis.}
\format{
A list containing one data frame and several vectors:
\itemize{
\item \code{ratings:} A 5 by 26 data frame with
the original ratings performed using a
5 point Likert rating scale with
the beers being rows and the evaluations
by the judges being the columns.
\item \code{judgeVector: }
A vector indicating what judge was rating what variable.
A 5 by 26 data frame with
\item \code{Var4Judges: } a vector storing
the number of variables per judge.
\item \code{namesOfJudges: } A vector storing
the names of the judges.
\item \code{color4Products: } A vector storing
the color names for the beers.
\item \code{color4Judges: } A vector storing
the color names for the judges.
}
}
\usage{
data("fiveBeers5Judges")
}
\description{
\code{fiveBeers5Judges}:
Five tasters evaluated
five beers on dimensions
that they had previously chosen.
These data can be used to illustrate,
among others,
Multiple Factor Analysis (MFA),
STATIS,  or Distatis.

The judges are named \code{J1}
to \code{J5}
}
\references{
These data were used in a couple of workshops
for the SPISE meeting.
}
\author{
Hervé Abdi
}
\keyword{data4PCCAR}
\keyword{datasets}
