% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveWines4Rotation.R
\docType{data}
\name{fiveWines4Rotation}
\alias{fiveWines4Rotation}
\title{5 wines are described by
Sensory and Chemistry variables.}
\format{
A list with two matrices:
\itemize{
\item \code{Xmat.Chemistry} The chemistry properties (4 variables)
\item \code{Ymat.Sensory} The sensory properties (3 variables).
}
}
\usage{
data("fiveWines4Rotation")
}
\description{
\code{fiveWines4Rotation}:
wines are described by two blocks of variables:
Sensory and Chemistry.
}
\details{
The five (rather fictitious) wines
are described by two blocks
of variables corresponding (respectively) to
their  sensory and chemistry properties.
This data set has been used to illustrate (Varimax) rotation
for factor analytic methods and also in several papers
about Partial Least Square Regression.
}
\references{
These data have been used to illustrate
rotations in factor analysis and PLSR. See, e.g.,
(papers available from
\url{https://personal.utdallas.edu/~herve/}):

Abdi, H. (2010). Partial least square regression,
projection on latent structure regression, PLS-Regression.
\emph{Wiley Interdisciplinary Reviews:
Computational Statistics, 2}, 97-106.

Abdi, H. (2003). Factor rotations.
In M. Lewis-Beck, A. Bryman, T. Futing (Eds):
\emph{Encyclopedia for research methods for the social sciences}.
Thousand Oaks (CA): Sage. 978-982.
}
\author{
Hervé Abdi
}
\keyword{data4PCCAR}
\keyword{datasets}
