% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodInFrance.R
\docType{data}
\name{foodInFrance}
\alias{foodInFrance}
\title{How much typical French families spent on different types
of food in the 1950's.}
\format{
A list containing 2 data frames:
\itemize{
\item \code{df.active:} the data per se
\item \code{supplementary.variables:} Additional information
long and short names and factors to expliciely identify
social class and number of children.
}
}
\usage{
data("foodInFrance")
}
\description{
\code{foodInFrance}: How much typical French families spent for food
in the 1950's. The families are categorized by social class
(manual workers, employees, and executives)
and by number of children
(from 2 to 5).
Interestingly (for historians of economics), at this
time, \code{Wine} was considered a food.
}
\details{
The measurements give the amount in \emph{Francs}
spend by a typical (i.e., average) family on 7 items:
\code{ Bread, Vegetables, Fruits, Meat, Poultry, Milk, Wine}

Because the unit (i.e., Francs spent) is the same all accross the table
(and a Franc is a Franc no matter what you buy),
this example illustrates the
case when the data should \emph{not} be normalized when
performing a PCA.
}
\references{
The original data were published in a (now unavailable)
report:

Tabard, N. (1967). \emph{Les Condition de Vie des Familles}.
Paris: UNCAF-CREDOC.

The concatenated data are given in

Lebart, L., Morinreau,, A., & Fenelon, J.P. (1982).
\emph{Traitement des Données Statistiques}. Paris: Dunod.

These data are also used in:

Abdi, H, & Willaims, L.A. (2010).
Principal component analysis.
\emph{Wiley Interdisciplinary Reviews: Computational Statistics, 2},
433-459.

and in:

Hardle, W., K., & Simar, L.  (2015) \emph{Applied Multivariate
Statistical Analysis}. New York: Springer.
}
\author{
Hervé Abdi from Nicole Tabard via  Ludovic Lebart et al.
}
\keyword{data4PCCAR}
\keyword{datasets}
