% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_internal_functions.R
\name{genPDQ2}
\alias{genPDQ2}
\title{genPDQ2, authored by Derek Beaton. 
Cf. package GSVD on GitHub, soon available on the CRAN}
\usage{
genPDQ2(
  datain,
  M = NULL,
  W = NULL,
  is.mds = FALSE,
  decomp.approach = "svd",
  k = 0
)
}
\description{
genPDQ2, authored by Derek Beaton. 
Cf. package GSVD on GitHub, soon available on the CRAN
}
\keyword{internal}
