% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiplot4PCA.R
\name{getBiplotCoor}
\alias{getBiplotCoor}
\title{Compute the coordinates of variables
for a PCA computed with \code{ExPosition::epPCA()}.
Temporary version}
\usage{
getBiplotCoor(res.epPCA, varLength = NULL, axis1 = 1, axis2 = 2)
}
\arguments{
\item{res.epPCA}{the output of \code{ExPosition::epPCA()}}

\item{varLength}{PARAM_DESCRIPTION, (\code{Default: NULL})}

\item{axis1}{horizontal axis (\code{Default: 1})
(used to compute the constraints).}

\item{axis2}{vertical axis (\code{Default: 2})
(used to compute the constraints).}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
\code{getBiplotCoor}
Compute the coordinates of variables
for a PCA computed with \code{ExPosition::epPCA()}.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[prettyGraphs]{minmaxHelper}}
}
\author{
Hervé Abdi
}
