% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCtr.12_try2.R
\name{getImportantCtr}
\alias{getImportantCtr}
\title{Get
the items with important contributions for a factorial plane.}
\usage{
getImportantCtr(ctr, eig, axis1 = 1, axis2 = 2)
}
\arguments{
\item{ctr}{The contributions of the items (e.g.,
the output of \code{ExPosition}
such as \code{$ExPosition.Data$ci}).}

\item{eig}{the eigenvalues of the analysis
(e.g.,
the output of \code{ExPosition}
 such as \code{$ExPosition.Data$eigs}).}

\item{axis1}{the horizontal axis, (Default: 1).}

\item{axis2}{the vertical axis  (Default: 2).}
}
\value{
a list with
\itemize{
\item{ctr.12: }{The contributions of the items to the plane;}
\item{importantCtr.12: }{A logical vector identifying
the items whose contribution to the plane
is larger than the average;}
\item{importantCtr.1or2: }{A logical vector identifying
the items whose contribution to the plane is larger than
the  average, or
is larger than the average for at least one of the components.}
}
}
\description{
\code{getImportantCtr.12}: 
 Gets the items whose contribution is
important for a factorial plane.
\code{getImportantCtr.12} gives
the contributions of the items to the plane, the items
important for the plane, and the items important for the plane
or for at least one dimension of the plane.
}
\details{
\code{getImportantCtr.12} is mostly used to create clean
factorial maps by plotting only the important items.
\code{getImportantCtr.12}:
can be used for row or columns of PCA, CA, MCA, etc.
}
\author{
Hervé Abdi
}
