% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodies4MCA.R
\name{getVarNames}
\alias{getVarNames}
\title{Get the names of variables from the levels
of the columns in an MCA performed with \code{ExPosition::epMCA}.}
\usage{
getVarNames(labelsNames)
}
\arguments{
\item{labelsNames}{The name of the labels,
typically from the output of \code{ExPosition::epMCA}}
}
\value{
A list with 3 vectors of the  original factor names:
\code{variableNames}: The names of the variables,
\code{stripedNames}: The names of the levels of the variables
striped of their extensions,
\code{originalLabelsNames}: The original label names.
}
\description{
\code{getVarNames}
Get the names of variables from the levels
of the columns in an MCA performed with \code{ExPosition::epMCA}.
}
\details{
This is a helper function for
some other helper functions for \code{ExPosition::epMCA}.
}
\examples{
\dontrun{
if(interactive()){
 # an example from ExPosition
library(ExPosition)
data(mca.wine)
lesRes <- epMCA(mca.wine$data, graphs = FALSE)
nameOffactors <- getVarNames(rownames(lesRes$ExPosition.Data$fj) )
 }
}
}
\seealso{
\code{\link[tools]{fileutils}}
}
\author{
Hervé Abdi
}
