% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduatingInUSA.R
\docType{data}
\name{graduatingInUSA}
\alias{graduatingInUSA}
\title{Number of University Graduates in the USA from 1966 to 2001.}
\format{
A list containing 7 data frames:
\itemize{
\item \code{df.all} the grand contingency table
\item \code{menBachelor}: the contingency table for men bachelors
\item \code{menMaster:} the contingency table for men maters
\item \code{menPhd} the contingency table for men Ph.D.
\item \code{womenBachelor}: the contingency table for women bachelors
\item \code{womenMaster:} the contingency table for women maters
\item \code{womenPhd} the contingency table for women Ph.D.
}
}
\usage{
data("graduatingInUSA")
}
\description{
\code{graduatingInUSA}: A series of contingency tables
giving the number of graudates in the USA from 1966 to 2001.
The grand contingency tables gives all the graduates by fields
and by year. the partial tables subset the grand tables
by gender (women vs men) and by lvel of graduation
(bachelor, master, and PH.D.).
}
\references{
I could not find the exact source of these data but they came
from some official agency.
}
\author{
Hervé Abdi
}
\keyword{data4PCCAR}
\keyword{datasets}
