% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_internal_functions.R
\name{\%^\%}
\alias{\%^\%}
\title{power, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN}
\usage{
x \%^\% power
}
\description{
power, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN
}
\keyword{internal}
