% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.diagMat.R
\name{is.diagMat}
\alias{is.diagMat}
\title{Test if an object is a diagonal matrix}
\usage{
is.diagMat(DATA)
}
\arguments{
\item{DATA}{an object to be tested.}
}
\value{
a logical value \code{TRUE} if the object is a matrix.
}
\description{
\code{is.diagMat} Test
if an object is a diagonal matrix
}
\examples{
is.diagMat(c(1,2,3))

}
\author{
Hervé Abdi
}
