% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_internal_functions.R
\name{is.diagonal.matrix}
\alias{is.diagonal.matrix}
\title{is.diagonal.matrix, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN}
\usage{
is.diagonal.matrix(x, tol = .Machine$double.eps)
}
\description{
is.diagonal.matrix, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN
}
\keyword{internal}
