% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_internal_functions.R
\name{is.empty.matrix}
\alias{is.empty.matrix}
\title{is.empty.matrix, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN}
\usage{
is.empty.matrix(x, tol = .Machine$double.eps)
}
\description{
is.empty.matrix, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN
}
\keyword{internal}
