% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_internal_functions.R
\name{is.identical.matrix}
\alias{is.identical.matrix}
\title{is.identical.matrix, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN}
\usage{
is.identical.matrix(x, tol = .Machine$double.eps, round.digits = 12)
}
\description{
is.identical.matrix, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN
}
\keyword{internal}
