% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_internal_functions.R
\name{isDiagonal.matrix}
\alias{isDiagonal.matrix}
\title{sqrt.mat, authored by Derek Beaton.
Cf. ExPosition CRAN}
\usage{
isDiagonal.matrix(X)
}
\description{
sqrt.mat, authored by Derek Beaton.
Cf. ExPosition CRAN
}
\keyword{internal}
