% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastInferences4-eigen-CA.R
\name{malinvaudQ4CA.perm}
\alias{malinvaudQ4CA.perm}
\title{Compute the Malinvaud/Saporta test for
the omnibus and dimensions of a correspondence
analysis.}
\usage{
malinvaudQ4CA.perm(Data, LesEigPerm = NULL, ndigit4print = 4)
}
\arguments{
\item{Data}{A matrix suitable for
correspondence analysis
(i.e., with all non-negative elements).}

\item{LesEigPerm}{the permuted eigenvalues
as a \code{niteration} times rank of the
data  matrix.
if  \code{LesEigPerm} is \code{NULL}
(Default), \code{malinvaudQ4CA.perm}
will only compute the normal
(i.e., chi2 based) approximation.}

\item{ndigit4print}{(Default: 4),
number of significant
digits to use to report the results.}
}
\value{
A (self-explanatory)
table with the results of the test
}
\description{
\code{malinvaudQ4CA.perm}:
Computes the Malinvaud / Saporta test for
the omnibus and dimensions of a correspondence
analysis. \code{malinvaudQ4CA.perm} gives
the asymptotic Chi2 values and their associated
\emph{p}-value under the usual assumptions.
If provided with permuted
CA-eigenvalues, \code{malinvaudQ4CA.perm}
will report the \emph{p}-value obtained from
the permutation test.
}
\details{
DETAILS
}
\examples{
\dontrun{
set.seed(87) # set the seed
X <- matrix(round(runif(21)*20), ncol = 3) # good for CA
resMalin <- malinvaudQ4CA.perm(X)
}
}
\references{
The original work is described in a rather
hard to find paper (published in the proceeding
of a meeting) by Malinvaud:

Malinvaud, E. (1987).
Data Analysis in applied socio-economic statistics
with special considerations of correspondence analysis.
\emph{Marketing Science Conference Proceedings},
HEC-ISA, Jouy-en-Josas.

A synthesis of the method with additional information
can be found
in

Saporta (2011). \emph{Probabilité
et Analyse des Données (3rd Ed)}.
Technip, Paris. p. 209.
}
\author{
Hervé Abdi
}
