% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentaImageryOSIQ.R
\docType{data}
\name{mentalImageryOSIQ}
\alias{mentalImageryOSIQ}
\title{(simulated) results of 2,100 participants answering
a Likert scale like questionnaire about mental imagery
(called the Object-Spatial Imagery Questionnaire: OSIQ).}
\format{a list containing one
data frame (\code{mentalImageryOSIQ$OSIQ})
with 2,100 rows (observations)
and 30 columns (questions). The first letter
of the column denotes the type of imagery
(s for spatial and o for object). The questions are
in the same order as in the original publication
by Blajenkova et al. (2006). The first letter
of the name of a row can be \code{H, M, L}.
This letter indicates that the participants identified
themselves as having \code{H}igh or superior episodic memory,
\code{M}edium (when they did not mention high nor low), or
\code{L}ow episodic memory.}
\usage{
data("mentalImageryOSIQ")
}
\description{
\code{mentalImageryOSIQ} a data set with
(simulated) results of 2,100 participants answering
a Likert scale like questionnaire comprising
30 questions about mental imagery
(called the Object-Spatial Imagery Questionnaire: OSIQ).
Half of the questions concern mental imagery for object and
the other half concern mental imagery for spatial locations.
}
\details{
These data are simulated data that roughly match some real data.
Details about the questionnaire and for the actual questions
see the paper by Blajenkova et al. (2006).
The asnwers to the questions
were given on a Likert type scale from 1 to 5. One question
(\code{s27}) has been reversed coded so that a large number
indicates a good memory.

These data were created to mimic the structure of real data and
were created with
the function \code{\link{buildRandomImage4PCA}}.
}
\references{
The original questionnaire can be found in:

Blajenkova, O.,
Kozhevnikov, M., & Motes, M.A. (2006).
Object-spatial imagery: a new self-report imagery questionnaire.
\emph{Applied Cognitive Psychology, 20}, 239–263.

These (simulated) data have been also used in
Guillemot et al. (in press, 2019).
}
\author{
Hervé Abdi & Brian Levine
}
\keyword{data4PCCAR}
\keyword{datasets}
