% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastInferences4-eigen-CA.R
\name{multinomCV4CA}
\alias{multinomCV4CA}
\title{Compute the permuted and bootstrapped eigenvalues
of the correspondence analysis (CA) of a matrix suitable
for CA (i.e., a matrix with non negative elements).}
\usage{
multinomCV4CA(Data, niter = 1000)
}
\arguments{
\item{Data}{a matrix suitable
for CA (i.e., a matrix with non negative elements).}

\item{niter}{number of Bootstrapped/Permutations
(Default: \code{1000}).}
}
\value{
A list with two elements: 1)
\code{$Permuted.ValP}: The matrix of the
\code{niter} by \code{rank(Data)} permuted
eigenvalues of the data matrix \code{Data}, and
\code{$Bootstraped.ValP}: The matrix of the
\code{niter} by \code{rank(Data)} bootstrapped
eigenvalues of the data matrix \code{Data}.
}
\description{
\code{multinomCV4CA}:
a very fast routine that
computes the permuted and bootstrapped eigenvalues
of the correspondence analysis (CA) of a matrix suitable
for CA (i.e., a matrix with non negative elements).
}
\details{
\code{multinomCV4CA}
uses the multinomial distribution to
simulate bootstrap and permutation resampling
for a correspondence analysis.
}
\examples{
\dontrun{
set.seed(87) # set the seed
X <- matrix(round(runif(21)*20), ncol = 3) # good for CA
ResCV <- multinomCV4CA(X)
}
}
