% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameExPoDF.R
\name{nameExpositionResults}
\alias{nameExpositionResults}
\title{add dimension names to the results from \code{ExPosition}
or \code{InPosition}
(for use with \code{ggplot2}).}
\usage{
nameExpositionResults(resExpo, names4Dimensions = "Dimension ")
}
\arguments{
\item{resExpo}{the output of an analysis performed
by \code{ExPosition} with the functions:
\code{epCA, epPCA, epGPCA, epMCA, epMDS}.}

\item{names4Dimensions}{names for
the dimensions, Default: 'Dimension '}
}
\value{
the original results with their dimensions 
named or renamed.
}
\description{
\code{nameExpositionResults}:
adds dimension names to the results of analyses performed with 
\code{ExPosition} (column names are needed for plots
created with \code{ggplot2}).
}
\details{
\code{ExPosition} returns the results of its analysis in a list
of matrices
that comprises the factor scores, the contributions, 
the cosines (etc. see documentation for \code{\link{ExPosition}}).
The columns of these statistics 
do not have names and this creates
a problem when using \code{ggplot2} whose
syntax  requires names for variables in \code{aes()}.
}
\examples{
\dontrun{
if(interactive()){
# Get the data from package data4PCCAR
df <- twentyWines$df.active
resPCA <- epPCA(df, scale = FALSE, graphs = FALSE, k = 1)
resPCA.named <- nameExpositionResults(resPCA)
 }
}
}
\seealso{
ExPosition
}
\author{
Hervé Abdi
}
