% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS_jack_svds_HA.R
\name{normaliz}
\alias{normaliz}
\title{FUNCTION_TITLE}
\usage{
normaliz(F)
}
\arguments{
\item{F}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Hervé Abdi, Lei Xuan
}
