% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm4CCA.R
\name{perm4CCA}
\alias{perm4CCA}
\title{Title}
\usage{
perm4CCA(DATA1, DATA2, center1 = TRUE, center2 = TRUE,
  scale1 = "ss1", scale2 = "ss1", nIter = 1000, permType = "byMat",
  compact = FALSE)
}
\arguments{
\item{DATA1}{}

\item{DATA2}{}

\item{center1}{}

\item{center2}{}

\item{scale1}{}

\item{scale2}{}

\item{nIter}{}

\item{permType}{}

\item{compact}{}
}
\description{
Title
}
\examples{
\dontrun{}
}
