% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodies4MCA.R
\name{phi2CT}
\alias{phi2CT}
\title{A fast function to compute the phi2 for a Contingency table.}
\usage{
phi2CT(CT)
}
\arguments{
\item{CT}{A contingency table.}
}
\value{
The phi2 (coefficient of correlation)
associated with the Contingency table.
}
\description{
\code{phi2CT}
A fast function to compute the phi2 for a Contingency table.
}
\details{
phi2 is the coefficient of correlation for a
contingency table. It is equal to  the inertia of
the contingency tables divided by the
min(number of rows, number of columns) - 1.
The inertia is also equal to the Chi2 of the table divided
by its total.
Phi2 could also be derived from the function
\code{chisq.test}.
}
\examples{
phi2CT(matrix(c(19,15,7,14,4,21,5,22,8,5,9,21), nrow = 4) )
}
\seealso{
phi2Mat4BurtTable
}
