% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodies4MCA.R
\name{phi2Mat4BurtTable}
\alias{phi2Mat4BurtTable}
\title{Compute the phi2 (correlation) matrix
for data tables suitable for Multiple Correspondence Analysis
(MCA).}
\usage{
phi2Mat4BurtTable(DATA, make_data_nominal = TRUE)
}
\arguments{
\item{DATA}{A data table suitable for MCA.
Could be an \eqn{I} by \eqn{K} data frame / matrix
with columns being factors, or an already
\eqn{I} by \eqn{J}
expanded matrix with 0/1 (as created for example
with \code{ExPosition::makeNominalData}).}

\item{make_data_nominal}{(Default: TRUE)
when \code{TRUE} the data are factors,
otherwise, the data are already in 0/1 form.}
}
\value{
A list with
\enumerate{
\item \code{phi2.mat}: The phi2 correlation matrix
and
\item \code{burt.table}: the Burt table.
}
}
\description{
\code{phi2Mat4BurtTable}
Compute the phi2 (correlation) matrix
for data tables suitable for Multiple Correspondence Analysis
(MCA).
}
\details{
The phi2 matrix from \code{phi2Mat4BurtTable}
can be used to create correlation heat maps for MCA.
}
\examples{
\dontrun{
if(interactive()){
library(ExPosition)
data("mca.wine")
phi2Mat4BurtTable(mca.wine$data)
  }
}
}
