% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS_jack_svds_HA.R
\name{PLS4jack}
\alias{PLS4jack}
\title{FUNCTION_TITLE}
\usage{
PLS4jack(X, Y, xsup, nfactor)
}
\arguments{
\item{X}{PARAM_DESCRIPTION}

\item{Y}{PARAM_DESCRIPTION}

\item{xsup}{PARAM_DESCRIPTION}

\item{nfactor}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MASS]{ginv}} \code{\link{PLSR_SVD}}
}
\author{
Hervé Abdi, Lei Xuan
}
