% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS_jack_svds_HA.R
\name{PLSR_SVD}
\alias{PLSR_SVD}
\title{version of PLS regression using the Singular
Value Decomposition instead of the original NIPALS}
\usage{
PLSR_SVD(X, Y, nfactor)
}
\arguments{
\item{X}{The \eqn{N} observations by \eqn{I} variables
matrix (\strong{X}) of the predictors.}

\item{Y}{The \eqn{N} observations by \eqn{I} variables
matrix (\strong{Y}) to be predicted.}

\item{nfactor}{Number of factors (i.e., \emph{latent variables})
to be used for the predictions. Note that the solution in PLSR
is strongly dependentg on the number of factors to keep}
}
\value{
More needed here.
In the meantime look at the print function.
}
\description{
\code{PLSR_SVD}: A
version of PLS regression using the Singular
Value Decomposition (SVD) instead of the original NIPALS.
}
\details{
GOAL:

Compute the PLS regression coefficients/decomposition
\strong{X} = \strong{T} * \strong{P}' Y = \strong{T} * \strong{B} * \strong{C}'
=  \strong{X} * \strong{B}pls  \strong{X} and \strong{Y} being \eqn{Z}-scores\preformatted{                     **B** = diag(b)
}

Y = X * Bpls_star with X being augmented with a col of ones
and Y and X having their original units
Yjack is the jackknifed (LOO) estimation of Y
T'*T = I (NB normalization <> than SAS)
W'*W = I
C is unit normalized,
U, P are not normalized
For notations: see Abdi (2003, 2007, 2010),
available from \code{personal.utdallas.edu/~herve}
}
\examples{
\dontrun{
# Run the wine example from Abdi (2010)
data("fiveWines4Rotation")
Xmat <- fiveWines4Rotation$Xmat.Chemistry
Ymat <- fiveWines4Rotation$Ymat.Sensory
resPLSR <- PLSR_SVD(Xmat, Ymat, 3)
}
}
\references{
(see also www.utd.edu/~herve)
\enumerate{
\item Abdi, H. (2010).
Partial least square regression,
projection on latent structure regression, PLS-Regression.
Wiley Interdisciplinary Reviews: Computational Statistics,
2, 97-106.
\item Abdi, H. (2007).
Partial least square regression (PLS regression).
In N.J. Salkind (Ed.):
Encyclopedia of Measurement and Statistics.
Thousand Oaks (CA): Sage. pp. 740-744.
\item Abdi. H. (2003).
Partial least squares regression (PLS-regression).
In M. Lewis-Beck, A. Bryman, T. Futing (Eds):
Encyclopedia for research methods for the social sciences.
Thousand Oaks (CA): Sage. pp. 792-795.
}
}
\author{
Hervé Abdi, Lei Xuan
}
