% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcoholInEurope.R
\name{print.alcoholInEU}
\alias{print.alcoholInEU}
\title{Change the print function for the data set:
\code{alcoholInEU}
#'
Change the print function for the data set:
\code{alcoholInEU}.}
\usage{
\method{print}{alcoholInEU}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{alcoholInEU}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{alcoholInEU}
#'
Change the print function for the data set:
\code{alcoholInEU}.
}
\author{
Herve Abdi
}
\keyword{internal}
