% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindSortingWines.R
\name{print.blindSortingWines}
\alias{print.blindSortingWines}
\title{Change the print function for the data set:
\code{blindSortingWines}
#'
Change the print function for the data set:
\code{blindSortingWines}.}
\usage{
\method{print}{blindSortingWines}(x, ...)
}
\arguments{
\item{x}{a list:
the data set \code{blindSortingWines}.}

\item{...}{the rest.}
}
\description{
Change the print function for the data set:
\code{blindSortingWines}
#'
Change the print function for the data set:
\code{blindSortingWines}.
}
\author{
Hervé Abdi
}
\keyword{internal}
