% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferences4PLSC.R
\name{print.bootBrick.ij.eig4plsc}
\alias{print.bootBrick.ij.eig4plsc}
\title{Change the print function for class bootBrick.ij.eig4plsc}
\usage{
\method{print}{bootBrick.ij.eig4plsc}(x, ...)
}
\arguments{
\item{x}{a list: output of \code{Boot4PLSC}}

\item{...}{everything else for the function}
}
\description{
Change the print function for objects
 of the class \code{bootBrick.ij.eig4plsc}
 (output of \code{Boot4PLSC})
}
\author{
Hervé Abdi
}
