% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferencesMultinom4CA-3.R
\name{print.bootBrick.ij}
\alias{print.bootBrick.ij}
\title{Change the print function for class bootBrick.ij}
\usage{
\method{print}{bootBrick.ij}(x, ...)
}
\arguments{
\item{x}{a list: output of fastBoot4CA}

\item{...}{everything else for the function}
}
\description{
Change the print function for bootBrick.ij
(output of fastBoot4CA)
}
\author{
Herve Abdi
}
