% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferences4PLSC.R
\name{print.bootBrick.ij4plsc}
\alias{print.bootBrick.ij4plsc}
\title{Change the print function for class bootBrick.ij4plsc}
\usage{
\method{print}{bootBrick.ij4plsc}(x, ...)
}
\arguments{
\item{x}{a list: output of Boot4PLSC}

\item{...}{everything else for the function}
}
\description{
Change the print function for bootBrick.ij4plsc
(output of Boot4MultCA)
}
\author{
Herve Abdi
}
