% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMC.R
\name{print.bootEigen}
\alias{print.bootEigen}
\title{Change the print function for the class \code{bootEigen}.}
\usage{
\method{print}{bootEigen}(x, ...)
}
\arguments{
\item{x}{a list: output of boot.eigen (class: bootEigen)}

\item{...}{everything else for the functions}
}
\description{
\code{print.bootEigen} Change the print function for
objects ot the class \code{bootEigen}
(output of \code{boot.eigen}).
}
\author{
Hervé Abdi
}
