% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorsOfVowels.R
\name{print.colorsOfVowels}
\alias{print.colorsOfVowels}
\title{Change the print function for the data set:
\code{colorsOfVowels}}
\usage{
\method{print}{colorsOfVowels}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {colorsOfVowels}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{colorsOfVowels}.
}
\author{
Herve Abdi
}
\keyword{internal}
