% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiplot4PCA.R
\name{print.coor4biplot}
\alias{print.coor4biplot}
\title{Change the print function for coor4biplot}
\usage{
\method{print}{coor4biplot}(x, ...)
}
\arguments{
\item{x}{a list: output of \code{\link{getBiplotCoor}}}

\item{...}{everything else for the functions}
}
\description{
Change the print function for coor4biplot
}
\author{
Hervé Abdi
}
