% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCtr.12_try2.R
\name{print.ctr12}
\alias{print.ctr12}
\title{Change the print function for objects of class \code{ctr12}}
\usage{
\method{print}{ctr12}(x, ...)
}
\arguments{
\item{x}{a list: output of \code{getImportantCtr.12}.}

\item{...}{everything else for the function.}
}
\description{
Change the print function for objects of class \code{ctr12}
}
\author{
Hervé Abdi
}
