% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadInTheUSA.R
\name{print.deadUSA}
\alias{print.deadUSA}
\title{Change the print function for the data set:
\code{deadUSA}}
\usage{
\method{print}{deadUSA}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{deadUSA}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{deadUSA}.
}
\author{
Herve Abdi
}
