% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionsOfDepression.R
\name{print.dimDepression}
\alias{print.dimDepression}
\title{Change the print function for the data set:
\code{dimDepression}
#'
Change the print function for the data set:
\code{dimDepression}.}
\usage{
\method{print}{dimDepression}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{dimDepression}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{dimDepression}
#'
Change the print function for the data set:
\code{dimDepression}.
}
\author{
Herve Abdi
}
