% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epGPCA2.R
\name{print.epGPCA2}
\alias{print.epGPCA2}
\title{Change the print function for the data set:
\code{foodOfTheWorld}}
\usage{
\method{print}{epGPCA2}(x, ...)
}
\arguments{
\item{x}{a list
that contains items
from the \code{epGPCA2} class.}

\item{...}{the rest (i.e.,
inherited/passed arguments for
\code{S3} print method(s).)}
}
\description{
Change the print function for the data set:
\code{foodOfTheWorld}
}
\author{
Derek Beaton, Cerise Chin-Fatt, Herve Abdi
}
\keyword{internal}
