% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epVari.R
\name{print.epVari}
\alias{print.epVari}
\title{Change the print function for createFactorMap}
\usage{
\method{print}{epVari}(x, ...)
}
\arguments{
\item{x}{a list: output of createFactorMap}

\item{...}{everything else for the functions}
}
\description{
Change the print function for createFactorMap
}
\author{
Hervé Abdi
}
