% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplementary4Vari.R
\name{print.epVariSup.j}
\alias{print.epVariSup.j}
\title{Change the print function for epVariSup.j}
\usage{
\method{print}{epVariSup.j}(x, ...)
}
\arguments{
\item{x}{a list: output of supplementary4Vari}

\item{...}{everything else for the functions}
}
\description{
Change the print function for epVariSup.j
}
\author{
Hervé Abdi
}
