% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveBeers5Judges.R
\name{print.fiveBeers5Judges}
\alias{print.fiveBeers5Judges}
\title{Change the print function for the data set:
\code{fiveBeers5Judges}
#'
Change the print function for the data set:
\code{fiveBeers5Judges}.}
\usage{
\method{print}{fiveBeers5Judges}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{fiveBeers5Judges}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{fiveBeers5Judges}
#'
Change the print function for the data set:
\code{fiveBeers5Judges}.
}
\author{
Herve Abdi
}
