% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveBeersHedonics.R
\name{print.fiveBeersRank}
\alias{print.fiveBeersRank}
\title{Change the print function for the data set:
\code{fiveBeersRank}
#'
Change the print function for the data set:
\code{fiveBeersRank}.}
\usage{
\method{print}{fiveBeersRank}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{fiveBeersRank}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{fiveBeersRank}
#'
Change the print function for the data set:
\code{fiveBeersRank}.
}
\author{
Herve Abdi
}
\keyword{internal}
