% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveWines4Rotation.R
\name{print.fiveWines4Rotation}
\alias{print.fiveWines4Rotation}
\title{Change the print function for the data set:
\code{fiveWines4Rotation}
#'
Change the print function for the data set:
\code{fiveWines4Rotation}.}
\usage{
\method{print}{fiveWines4Rotation}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{fiveWines4Rotation}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{fiveWines4Rotation}
#'
Change the print function for the data set:
\code{fiveWines4Rotation}.
}
\author{
Hervé Abdi
}
\keyword{internal}
