% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodInFrance.R
\name{print.foodInFrance}
\alias{print.foodInFrance}
\title{Change the print function for the data set:
\code{foodInFrance}
#'
Change the print function for the data set:
\code{foodInFrance}.}
\usage{
\method{print}{foodInFrance}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {foodInFrance}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{foodInFrance}
#'
Change the print function for the data set:
\code{foodInFrance}.
}
\author{
Herve Abdi
}
