% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodOfTheWorld.R
\name{print.foodOfTheWorld}
\alias{print.foodOfTheWorld}
\title{Change the print function for the data set:
\code{foodOfTheWorld}}
\usage{
\method{print}{foodOfTheWorld}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {foodOfTheWorld}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{foodOfTheWorld}
}
\author{
Herve Abdi
}
