% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduatingInUSA.R
\name{print.graduatingUSA}
\alias{print.graduatingUSA}
\title{Change the print function for the data set:
\code{graduatingUSA}
#'
Change the print function for the data set:
\code{graduatingUSA}.}
\usage{
\method{print}{graduatingUSA}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{graduatingUSA}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{graduatingUSA}
#'
Change the print function for the data set:
\code{graduatingUSA}.
}
\author{
Herve Abdi
}
\keyword{internal}
