% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastInferences4-eigen-CA.R
\name{print.Inference4CA}
\alias{print.Inference4CA}
\title{Change the print function for Inference4CA}
\usage{
\method{print}{Inference4CA}(x, ...)
}
\arguments{
\item{x}{a list: output of, e.g.,  multinormCV4CA}

\item{...}{everything else for the functions}
}
\description{
Change the print function for Inference4CA
}
\author{
Hervé Abdi
}
