% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentaImageryOSIQ.R
\name{print.OSIQ}
\alias{print.OSIQ}
\title{Change the print function for the data set:
\code{OSIQ}
#'
Change the print function for the data set:
\code{OSIQ}.}
\usage{
\method{print}{OSIQ}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{OSIQ}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{OSIQ}
#'
Change the print function for the data set:
\code{OSIQ}.
}
\author{
Herve Abdi
}
