% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferencesMultinom4CA-3.R
\name{print.perm4CA}
\alias{print.perm4CA}
\title{Change the print function for perm4CA class}
\usage{
\method{print}{perm4CA}(x, ...)
}
\arguments{
\item{x}{a list: output of perm4RowCA}

\item{...}{everything else for the functions}
}
\description{
Change the print function for perm4CA class
objects
(output of astPerm4CA)
}
\author{
Hervé Abdi
}
