% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMC.R
\name{print.randEigen}
\alias{print.randEigen}
\title{Change the print function for the class \code{randEigen}.}
\usage{
\method{print}{randEigen}(x, ...)
}
\arguments{
\item{x}{a list: output of monteCarlo.eigen (class: randEigen)}

\item{...}{everything else for the functions}
}
\description{
\code{randEigen} Change the print function for
object of the class \code{randEigen} (ouput of monteCarlo.eigen).
}
\author{
Hervé Abdi
}
