% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixAuthorsPunctuated.R
\name{print.sixAuthors}
\alias{print.sixAuthors}
\title{Change the print function for the data set:
\code{sixAuthors}
#'
Change the print function for the data set:
\code{sixAuthors}.}
\usage{
\method{print}{sixAuthors}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {sixAuthors}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{sixAuthors}
#'
Change the print function for the data set:
\code{sixAuthors}.
}
\author{
Herve Abdi
}
