% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixBeers12Descriptors10Judges.R
\name{print.sixBeers12Descriptors}
\alias{print.sixBeers12Descriptors}
\title{Change the print function for the data set:
\code{sixBeers12Descriptors}
#'
Change the print function for the data set:
\code{sixBeers12Descriptors}.}
\usage{
\method{print}{sixBeers12Descriptors}(x, ...)
}
\arguments{
\item{x}{a list: the data set
\code{sixBeers12Descriptors}.}

\item{...}{the rest.}
}
\description{
Change the print function for the data set:
\code{sixBeers12Descriptors}
#'
Change the print function for the data set:
\code{sixBeers12Descriptors}.
}
\author{
Hervé Abdi
}
