% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixFaces.R
\name{print.sixFaces}
\alias{print.sixFaces}
\title{Change the print function for the data set:
\code{sixFaces}}
\usage{
\method{print}{sixFaces}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{sixFaces}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{sixFaces}.
}
\author{
Herve Abdi
}
\keyword{internal}
