% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixteenGums4Descriptors.R
\name{print.sixteenGums}
\alias{print.sixteenGums}
\title{Change the print function for the data set:
\code{sixteenGums}
#'
Change the print function for the data set:
\code{sixteenGums}.}
\usage{
\method{print}{sixteenGums}(x, ...)
}
\arguments{
\item{x}{a list: the data set
\code{sixteenGums}.}

\item{...}{the rest.}
}
\description{
Change the print function for the data set:
\code{sixteenGums}
#'
Change the print function for the data set:
\code{sixteenGums}.
}
\author{
Hervé Abdi
}
