% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOfMusic.R
\name{print.str_colorsOfMusic}
\alias{print.str_colorsOfMusic}
\title{Change the print function for the data set: 'str_colorsOfMusic'}
\usage{
\method{print}{str_colorsOfMusic}(x, ...)
}
\arguments{
\item{x}{a list: the data set: 'str_colorsOfMusic'}

\item{...}{the rest}
}
\description{
Change the print function for the data set: 'str_colorsOfMusic'
}
\author{
Hervé Abdi
}
\keyword{internal}
