% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary4PLSC.R
\name{print.supElementsPLS}
\alias{print.supElementsPLS}
\title{Change the print function for supElementsPLS}
\usage{
\method{print}{supElementsPLS}(x, ...)
}
\arguments{
\item{x}{a list: output of supplementary4Vari}

\item{...}{everything else for the functions}
}
\description{
Change the print function for supElementsPLS
}
\author{
Hervé Abdi
}
