% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementaryVariables4PLSC.R
\name{print.supVar4PLSC}
\alias{print.supVar4PLSC}
\title{Change the print function for supVar4PLSC}
\usage{
\method{print}{supVar4PLSC}(x, ...)
}
\arguments{
\item{x}{a list: output of supplementaryVariables4PLSC}

\item{...}{everything else for the functions}
}
\description{
Change the print function for supVar4PLSC
}
\author{
Hervé Abdi
}
