% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twentyWines.R
\name{print.twentyWines}
\alias{print.twentyWines}
\title{Change the print function for the data set:
\code{twentyWines}
#'
Change the print function for the data set:
\code{twentyWines}.}
\usage{
\method{print}{twentyWines}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {twentyWines}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{twentyWines}
#'
Change the print function for the data set:
\code{twentyWines}.
}
\author{
Herve Abdi
}
