% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winesOf3Colors.R
\name{print.wines3Colors}
\alias{print.wines3Colors}
\title{Change the print function for the data set:
\code{wines3Colors}
#'
Change the print function for the data set:
\code{wines3Colors}.}
\usage{
\method{print}{wines3Colors}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {wines3Colors}}

\item{...}{the rest}
}
\description{
Change the print function for the data set:
\code{wines3Colors}
#'
Change the print function for the data set:
\code{wines3Colors}.
}
\author{
Herve Abdi
}
