% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary4PLSC.R
\name{projSupplementaryObservations4PLSC}
\alias{projSupplementaryObservations4PLSC}
\title{Compute latent variables / factor scores for supplementary
observations in a \code{PLSC} analysis 
computed from \code{TExPosition::tepPLS}.}
\usage{
projSupplementaryObservations4PLSC(
  mat4sup,
  center_fixed,
  scale_fixed,
  loadings_fixed
)
}
\arguments{
\item{mat4sup}{The observations to be projected}

\item{center_fixed}{The barycenter vector (i.e., mean)
used to center the original variables.}

\item{scale_fixed}{the normlizing vector
used to scale the original variables.}

\item{loadings_fixed}{the loadings 
used to project the supplementary observations on the 
variables (obtaiined from the original analysis).}
}
\value{
the supplementary latent variable values.
}
\description{
\code{projSupplementaryObservations4PLSC}
computes latent variables / factor scores for supplementary
observations in a \code{PLSC} analysis 
computed from \code{TExPosition::tepPLS}. 
\code{projSupplementaryObservations4PLSC}
preprocesses and projects supplementary observations for a PLSC
by first preprocessing the variables (i.e., centering and scaling)
and then projecting the preprocess observations 
onto the loadings (a.k.a., saliences) from the original analysis.
 \code{projSupplementaryObservations4PLSC} is mostly used
 as a preparatory step for
  \code{supplementaryObservations4PLSC}.
}
\details{
\code{projSupplementaryObservations4PLSC}
is mostly a subroutine of
\code{\link{supplementaryObservations4PLSC}}.
}
\seealso{
\code{\link[TExPosition]{tepPLS}} 
\code{\link{supplementaryObservations4PLSC}}
}
\author{
Hervé Abdi
}
