% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenormInertiaExpo.R
\name{renormInertiaExPo}
\alias{renormInertiaExPo}
\title{Renorm the output of an \code{ExPosition analysis}.}
\usage{
renormInertiaExPo(resExpo, newNorm = NULL)
}
\arguments{
\item{resExpo}{the output of an n\code{ExPosition} function (e.g.,
\code{epCA, epPCA, epMCA}).}

\item{newNorm}{The new norm (Default: \code{NULL}),
if \code{NULL} the results are normed to a value of 1,
if \code{newNorm} is a scalar, the data will be nromed}
}
\value{
The renormed \code{ExPosition} output.
}
\description{
\code{renormInertiaExPo}: renorms
the output of an \code{ExPosition} function (e.g.,
\code{epCA, epPCA, epMCA}) to a given constant.
After normalization, all the objects whose inertia
depends upon the data will now have their inertia
equal to the new specified inertia.

The normalization constant can be a scalar, or
can be the output of a previous
\code{ExPosition}
analysis, in this last case
the inertia is the inertia
of the table used in this previous analysis.
}
\details{
All the dataframes in \code{resExpo}
whose norm depends upon the data
(i.e., \code{fi, fj, di, dj, pdq$Dv, pdq$Dd, eigs})
are renormed in the output.
}
\examples{
\dontrun{
data("mtcars") # use the mtcars data set 
resPCA.normed  <- renormInertiaExPo(epPCA(mtcars, graphs = FALSE), ncol(mtcars))
}
}
\seealso{
\code{\link{firstpos}}, \code{\link{firstpos4ExPo}}
}
\author{
Hervé Abdi
}
