% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMC.R
\name{scale0}
\alias{scale0}
\title{center and normalize a vector to 1 or to (\eqn{N} - 1)}
\usage{
scale0(x, center = TRUE, scale = "SS1")
}
\arguments{
\item{x}{a vector with \eqn{N} elements.}

\item{center}{if \code{TRUE} (Default) center if not do nothing.}

\item{scale}{if \code{FALSE} do nothing,
if \code{TRUE} normalize like \eqn{Z}-scores
[i.e., norm = \eqn{N} - 1], if \code{'SS1'} (Default) normalise
to 1 (i.e., norm = 1).}
}
\value{
a scaled version of \code{x}.
}
\description{
center and normalize a vector to 1 or to (\eqn{N} - 1)
}
\details{
When \code{x} comprises different values
\code{scale0} gives the same results as \code{scale}
except for the option \code{'SS1'}; when \code{x}
comprises all identical values and when the
parmater \code{scale = TRUE}, \code{scale} gives back
a vector of \code{NaN} (because of a division by 0 error),
whereas \code{scale0} gives back a vector of \code{0}.
}
\examples{
y <- scale0(c(1,2,5,10))

}
\seealso{
scale
}
\author{
Hervé Abdi
}
