% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixAuthorsPunctuated.R
\docType{data}
\name{sixAuthorsPunctuated}
\alias{sixAuthorsPunctuated}
\title{The punctuation used by six classical French Writers}
\format{A list containing 3 data frames:
\itemize{
\item \code{df.active:} the punctuation by the six authors
\item \code{supplementary.variables:} gives the details of the punctuation
marks included the \code{Other} category:
' \code{Exclam, Question, SemiCol, Colon}
\item \code{supplementary.observation:}
The punctuation marks used in
Chapter 1 of \emph{Les Réseaux de Neurones} written
by Hervé Abdi
}}
\usage{
data("sixAuthorsPunctuated")
}
\description{
\code{sixAuthorsPunctuated}:
The punctuation marks (comma, period, and "other marks")
used by six French authors Rousseau,
Chateaubrian, Hugo, Zola, Proust, Giraudoux.
This set is used to illustrate Correspondence Analysis.
}
\details{
The punctuation by the six authors constitutes the "active set"
in correspondence analysis: it is stored
in the data frame: \code{df.active}.

The category \code{Other} is the sum of 4 punctation marks:
\code{Exclam, Question, SemiCol, Colon} whose distribution
is stored in the data frame   \code{supplementary.variables}.

To illustrate the projection of supplementary onservations,
the data set also contains the punctutation from
Chapter 1 of \emph{Les Réseaux de Neurones} written
by Hervé Abdi. These data are stored in the data frame:
\code{supplementary.observation}.
}
\references{
The active data and supplementary variables come form:

Brunet E. (1989). Faut-il pondérer les données linguistiques?.
\emph{CUMFID}, 1989, 16:39–50.

These data are used in several papers illustrating
correspondence analysis such as:

Abdi H, Williams LJ. (2010). Correspondence analysis. In:
Salkind NJ, ed.
\emph{Encyclopedia  of  Research  Design}
Thousand Oaks: Sage Publications.
}
\author{
Hervé Abdi from E. Brunet
}
\keyword{data4PCCAR}
\keyword{datasets}
