% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixBeers12Descriptors10Judges.R
\docType{data}
\name{sixBeers12Descriptors10Judges}
\alias{sixBeers12Descriptors10Judges}
\title{Ten tasters evaluated
the intensity of twelve descriptors
for
six beers.
These data can be used to illustrate:
un-normed  (or normed)
Principal Component Analysis (PCA)
or Correspondence Analysis (CA).}
\format{
A list
(of class: \code{sixBeers12Descriptors})
containing
one data frame,
one tibble, and several vectors:
\itemize{
\item \code{ratingsIntensity: }
A data frame of dimension
\eqn{I = }6 rows (Beers) by \eqn{J = }12 columns
(descriptors) with
the average intensity ratings  performed using a
0 to 7  Likert rating scale.
\item \code{tb: } A tibble with the same
data as the data frame \code{ratingsIntensity},
but
with in addition three (character) columns describing
the beers (see below \code{shortNames},
\code{brewedIn}, \code{color4I}).
\item \code{longNamesBeers: }
A 6 element vector with the full names of the beers.
\code{brewedIn},
A 6 element vector with the name
of the country of origin of the
beers.
\item \code{color4Products: }
A 6 element vector with color names matching the
origin of the beers.
\item \code{descripteurs: }
A
12 element vector with the French name of the
descriptors.
\item \code{color4Descriptors: }
A
12 element vector with color names for the
descriptors.
}
}
\usage{
data("sixBeers12Descriptors10Judges", 
       package = 'data4PCCAR')
}
\description{
\code{sixBeers12Descriptors10Judges}:
tasters evaluated
the intensity of twelve descriptors
for
six beers.
These data can be used to illustrate:
un-normed  (or normed)
Principal Component Analysis (PCA)
and Correspondence Analysis (CA).
The data are the averages
(computed
over the 10 judges) of the
intensity ratings of the descriptors for
the beers.
}
\references{
These data are used in
Abdi H, Gomez, C., & Delmas, M. (2022).
Méthodes Statistiques Multivariées pour l'Analyse
Sensorielle et les Etudes Consommateurs.
}
\author{
Hervé Abdi & Carlos Gomez.
}
\keyword{data4PCCAR}
\keyword{datasets}
