% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixteenGums4Descriptors.R
\docType{data}
\name{sixteenGums4Descriptors}
\alias{sixteenGums4Descriptors}
\title{the average data
from ten tasters who evaluated
the intensity of four descriptors
for
sixteen chewing-gums.
These data can be used to illustrate:
normed
Principal Component Analysis (PCA)
or Multiple Correspondence Analysis (MCA).}
\format{
A list
(of class: \code{sixteenGums})
containing
one data frame,
and several vectors:
\itemize{
\item \code{ratingsIntensity: }
A data frame of dimension
\eqn{I = }16 rows (Beers) by
\eqn{J = }4 columns
(descriptors) with
the average intensity ratings  performed using
different scales per variables
(respectively: \code{10, 5, 5, 100} points).
\item \code{longNames: }
a 16 element vector with the long names of the gums.
\code{color4Products}
a 16 element vector with the
color names of the gums.
\code{color4Descriptors}
a 4 element vector with the
color names of the descriptors.
\code{scale}
a 4 element vector with the
number of points of the scales used
to describe the gums.
}
}
\usage{
data("sixteenGums4Descriptors", 
       package = 'data4PCCAR')
}
\description{
\code{sixteenGums4Descriptors}:
the average data obtained
from ten tasters who evaluated
sixteen chewing gums on
the intensity of four descriptors.
These data can be used to illustrate:
un-normed  (or normed)
Principal Component Analysis (PCA)
and Correspondence Analysis (MCA).
The data are the averages
(computed
over the 10 judges) of the
intensity ratings of the 4 descriptors for
the 16 gums.
}
\references{
These data are used in
Abdi H, Gomez, C., & Delmas, M. (2022).
Méthodes Statistiques Multivariées pour l'Analyse
Sensorielle et les Etudes Consommateurs.
}
\author{
Hervé Abdi & Carlos Gomez.
}
\keyword{data4PCCAR}
\keyword{datasets}
