% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementaryVariables4PLSC.R
\name{supplementaryVariables4PLSC}
\alias{supplementaryVariables4PLSC}
\title{Project supplementary variables (columns)
for a PLSC analysis (from  \code{\link[TExPosition]{tepPLS}}).}
\usage{
supplementaryVariables4PLSC(
  var.sup,
  resPLSC,
  Xset = NULL,
  Yset = NULL,
  X.center = TRUE,
  X.scale = "SS1",
  Y.center = TRUE,
  Y.scale = "SS1",
  dimNames = "Dimension "
)
}
\arguments{
\item{var.sup}{\strong{V}sup: The \eqn{N} by \eqn{K}
matrix of \eqn{K} supplementary
variables.}

\item{resPLSC}{the results of
a PLSC analysis performed with
\code{\link[TExPosition]{tepPLS}}.}

\item{Xset}{the original \strong{X} (\eqn{N} by \eqn{I})
data matrix. If \code{NULL}, the supplementary data
are projected on the dual set (i.e., \strong{Y}).
See also \code{details} for more.}

\item{Yset}{the original \strong{Y} (\eqn{N} by \eqn{J})
data matrix. If \code{NULL}, the supplementary data
are projected on the dual set (i.e., \strong{X}).
See also \code{details} for more.}

\item{X.center}{centering parameter for \strong{X}
(Default: \code{TRUE}).
See \code{\link[ExPosition]{expo.scale}} for details.}

\item{X.scale}{scaling parameter for \strong{X}
(Default: \code{'SS1'}).
See \code{\link[ExPosition]{expo.scale}} for details.}

\item{Y.center}{centering parameter for \strong{Y}
(Default: \code{TRUE}).
See \code{\link[ExPosition]{expo.scale}} for details.}

\item{Y.scale}{scaling parameter for \strong{Y}
(Default: \code{'SS1'}).
See \code{\link[ExPosition]{expo.scale}} for details.}

\item{dimNames}{Names for the
dimensions (i.e., factors) for the
supplementary loadings (Default: \code{'Dimension '}).}
}
\value{
a list with the following elements:
\itemize{
\item{"\code{cor.lx}": }{The correlations between the supplementary variables
and the \strong{X} set.}
\item{"\code{cor.ly}": }{The correlations between the supplementary variables
and the \strong{Y} set.}
\item{"\code{loadings.sup.X}": }{The loadings of the supplementary variables
as originating from the \code{Xset} (needs to
have the dual \code{Yset} to be computed).}
\item{"\code{loadings.sup.Y}": }{The loadings of the supplementary variables
as originating from the \code{Yset} (needs to
have the dual \code{Xset} to be computed).}
}
}
\description{
\code{supplementaryVariables4PLSC}:
Projects supplementary variables (columns)
for a PLSC analysis
(from  \code{\link[TExPosition]{tepPLS}}).
The variables should be measured on the same observations
as the observations measured on the original analysis.
The original data consisted in 2 matrices denoted
\strong{X} (dimensions \eqn{N} by \eqn{I})  and
\strong{Y} (\eqn{N} by \eqn{J}).
The supplementary data denoted \strong{V}sup is a
\eqn{N} by \eqn{K} matrix, that can be considered
as originating either from \strong{X}
(and then denoted \strong{X}sup) or \strong{Y}
(and then denoted \strong{Y}sup) .
If originating from \strong{X}
(resp, \strong{Y}) matrix \strong{Y} (resp, \strong{X})
is the \emph{dual} matrix.
Note that \emph{only the dual matrix}
is needed to project supplementary
variables.
See \code{details} for more.
}
\details{
The computation relies on the SVD
of the correlation matrix between \strong{X} and \strong{Y},
computed as \strong{R} = \strong{X}'\strong{Y}
(where \strong{X} \strong{Y} are the original data matrices that
have been preprocessed,
with, e.g., centering and scaling)
and decomposed with the
SVD as \strong{R} = \strong{PDQ}', with the usual constraints that
\strong{P}'\strong{P} = \strong{Q}'\strong{Q} = \strong{I}.
\subsection{Active loadings}{

The active loadings are \strong{P} (\strong{X}-loadings)
and \strong{Q} (\strong{Y}-loadings). These loadings come
from the SVD of matrix \strong{R} = \strong{X}'\strong{Y} = \strong{PDQ}'.
}

\subsection{Transition formulas}{

The loadings of one set can be obtained from the correlation matrix
\strong{R} and the loadings from the dual set. For example:

\strong{P} = \strong{X}'\strong{YQ} inv(\strong{D}) =  \strong{RQ} inv(\strong{D})
(with inv(\strong{D}) being the inverse of \strong{D}). Eq. 1
}

\subsection{Projection of supplementary variables}{

Supplementary variable loadings are obtained by first computing
their correlation with their dual set and then projecting these
on the singular vector of their dual set. So, for example,
the loadings denoted \strong{P}sup for
an  \eqn{N}  by \eqn{K} matrix of
\eqn{K}
supplementary variables considered
as belonging to the \strong{X}-set will be projected by first computing
the correlation matrix between these variables and
\strong{Y} (the dual set) as: \strong{R}sup = \strong{X}sup' \strong{Y} (note that
we assume here that \strong{X}sup has been pre-processed in the same way as
\strong{X}). The supplementary loadings are now computed
by replacing in Eq.1 \strong{X} by \strong{X}sup to obtain:

\strong{P}sup = \strong{X}sup' * \strong{Y} * \strong{Q} * inv(\strong{D})
=  \strong{R}sup * \strong{Q} *  inv(\strong{D})
. Eq.2.
}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\references{
See:

Abdi, H., & Williams, L.J. (2013). Partial least squares methods:
Partial least squares correlation
and partial least square regression.
In: B. Reisfeld & A. Mayeno (Eds.),
\emph{Methods in Molecular Biology:
Computational Toxicology}. New York: Springer Verlag.
pp. 549-579.

Abdi, H. (2007). Singular Value Decomposition (SVD)
and Generalized Singular Value Decomposition (GSVD).
In N.J. Salkind (Ed.):
\emph{Encyclopedia of Measurement and Statistics}.
Thousand Oaks (CA): Sage. pp. 907-912.
}
\seealso{
\code{\link[TExPosition]{tepPLS}}
\code{\link[ExPosition]{expo.scale}}
\code{\link{supplementaryObservations4PLSC}}
}
\author{
Hervé Abdi
}
