% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementaryVariables4PLSCA.R
\name{supplementaryVariables4PLSCA}
\alias{supplementaryVariables4PLSCA}
\title{Project supplementary variables (columns)
for a PLSCA analysis (from  \code{\link[TExPosition]{tepPLSCA}}).
**Beta Version. Current Version 07/30/2020. **}
\usage{
supplementaryVariables4PLSCA(
  var.sup,
  make.var.sup.nominal = TRUE,
  resPLSCA,
  Xset = NULL,
  make.Xset.nominal = TRUE,
  Yset = NULL,
  make.Yset.nominal = TRUE,
  dimNames = "Dimension "
)
}
\arguments{
\item{var.sup}{\strong{V}sup: The \eqn{N} by \eqn{K}
matrix of \eqn{K} supplementary
variables.}

\item{make.var.sup.nominal}{logical, when  \code{TRUE}) (default)
transforms each column of \code{var.sup}
from a factor with \eqn{M} levels
to a set of \code{M} 0/1 vectors (to create a group coding,
also called \emph{complete disjonctive coding}).}

\item{resPLSCA}{the results of
a PLSC analysis performed with
\code{\link[TExPosition]{tepPLSCA}}.}

\item{Xset}{the original \strong{X} (\eqn{N} by \eqn{I})
data matrix. If \code{NULL}, the supplementary data
are projected on the dual set (i.e., \strong{Y}).
See also \code{details} for more.}

\item{make.Xset.nominal}{logical, when  \code{TRUE}) (default)
transforms each column of
\code{Xset} from a factor with \eqn{M} levels
to a set of \code{M} 0/1 vectors (to create a group coding,
also called \emph{complete disjonctive coding}).}

\item{Yset}{the original \strong{Y} (\eqn{N} by \eqn{J})
data matrix. If \code{NULL}, the supplementary data
are projected on the dual set (i.e., \strong{X}).
See also \code{details} for more.}

\item{make.Yset.nominal}{logical, when  \code{TRUE}) (default)
transforms each column of \code{Yset}
from a factor with \eqn{M} levels
to a set of \code{M} 0/1 vectors (to create a group coding,
also called \emph{complete disjonctive coding}).}

\item{dimNames}{Names for the
dimensions (i.e., factors) for the
supplementary loadings (Default: \code{'Dimension '}).}
}
\value{
a list with the following elements:
\itemize{

\item{"\code{loadings.sup.X}": }{The loadings
of the supplementary variables
as originating from the \code{Xset} (needs to
have the dual \code{Yset} to be computed).}
\item{"\code{sup.fi}": }{The singular-value-normalized
loadings of the supplementary variables
as originating from the \code{Xset} (needs to
have the dual \code{Yset} to be computed).}
\item{"\code{loadings.sup.Y}": }{The loadings
of the supplementary variables
as originating from the \code{Yset} (needs to
have the dual \code{Xset} to be computed).}
\item{"\code{sup.fj}": }{The singular-value-normalized
loadings of the supplementary variables
as originating from the \code{Yset} (needs to
have the dual \code{Xset} to be computed).}
\item{"\code{cor.lx}": }{The correlations
between the supplementary variables
and the \strong{X} set.}
\item{"\code{cor.ly}": }{The correlations
between the supplementary variables
and the \strong{Y} set.}
}
}
\description{
\code{supplementaryVariables4PLSCA}:
Projects supplementary variables (columns)
for a \code{PLSCA} analysis
(from  \code{\link[TExPosition]{tepPLSCA}}).
The variables should be measured on the same observations
as the observations used in the original analysis.
The original data consisted in 2
matrices (containing non-negative numbers such as count, as in
correspondece analysis,
or often simply 0/1 as in multiple correspondence analysis) denoted
\strong{X} (dimensions \eqn{N} by \eqn{I})  and
\strong{Y} (\eqn{N} by \eqn{J}).
The supplementary data denoted \strong{V}sup is a
\eqn{N} by \eqn{K} matrix, that can be considered
as originating either from \strong{X}
(and then denoted \strong{X}sup) or \strong{Y}
(and then denoted \strong{Y}sup) .
If originating from \strong{X}
(resp, \strong{Y}) matrix \strong{Y} (resp, \strong{X})
is the \emph{dual} matrix.
Note that \emph{only the dual matrix}
is needed to project supplementary
variables.
See \code{details} for more.
}
\details{
The computation relies on the Generalized
singular decomposition (GSVD)
of the contingency between \strong{X} and \strong{Y},
computed as \strong{R} = \strong{X}'\strong{Y}
(where \strong{X} \strong{Y} are the original data matrices that
have been preprocessed as for the original \code{PLSCA} analysis,
e.g., transformed into 0/1 vectors)
and decomposed with the
GSVD as \strong{R} = \strong{PDQ}', with the (metrics) constraints that
\strong{P}'_inv(\strong{Dr}) \strong{P} = \strong{Q}'inv(\strong{Dc})\strong{Q} = \strong{I} where
inv() denotes the inverse matrix and where \strong{Dr} (resp \strong{Dc})
are the diagonal matrices of the barycenters of (respectively)
\strong{X} and \strong{Y}.
\subsection{Transition formulas}{

The loadings of one set can be obtained
from the cross-product matrix
\strong{R} and the loadings from the dual set. For example:
if we denote \strong{Delta} the diagona; matrix of
the singular values,
\strong{F}  (resp. \strong{G}) the singular value normalized
factor scores (denoted \code{fi}, resp. \code{fj},
in \code{PLSCA}),
and \strong{L} (resp. \strong{C}) the row (resp. column profiles) the
the loadings of one set are derived from the other set as:

\strong{F} = \strong{LG} inv(\strong{Delta})  and
\strong{G} = \strong{CF} inv(\strong{Delta})
(with inv(\strong{Dc}) being the inverse of \strong{Dc}). Eq. 1
}

\subsection{Projection of supplementary variables}{

Supplementary variable loadings are obtained by first computing
the cross-product matrix with their dual set
and then using the transition formulas from
correspondence analysis to
compute one set of loadings from the loadings of the other set.
So, for example the loadings denoted \code{fii} for the
variables of the \code{Xset} can be obtained from the
row profiles of the \strong{R}sup matrix by replacing in Eq.1
\strong{L} by \strong{L}sup.
}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\references{
See:

Beaton, D., Dunlop, J., ADNI, & Abdi, H. (2016).
Partial Least Squares-Correspondence Analysis (PLSCA):
A framework to simultaneously analyze behavioral and genetic data.
\emph{Psychological Methods, 21}, 621-651.

Abdi H. & Béra, M. (2018).
Correspondence analysis.
In R. Alhajj and J. Rokne (Eds.),
\emph{Encyclopedia of Social Networks and Mining (2nd Edition)}.
New York: Springer Verlag.

Abdi, H. (2007). Singular Value Decomposition (SVD)
and Generalized Singular Value Decomposition (GSVD).
In N.J. Salkind (Ed.):
\emph{Encyclopedia of Measurement and Statistics}.
Thousand Oaks (CA): Sage. pp. 907-912.
}
\seealso{
\code{\link[TExPosition]{tepPLSCA}}
\code{\link[ExPosition]{makeNominalData}}
\code{\link[ExPosition]{makeRowProfiles}}
\code{\link{supplementaryObservations4PLSC}}
}
\author{
Hervé Abdi
}
