% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferences4PLSC.R
\name{sv2}
\alias{sv2}
\title{Compute the squared singular values of a matrix.}
\usage{
sv2(X)
}
\arguments{
\item{X}{a rectangular matrix (or dataframe)}
}
\value{
a vector of the squared singular values.
}
\description{
\code{sv2}: computes the squared singular values
(i.e., eigen-values)
of a matrix.
}
\details{
\code{sv2} is wraper around \code{eigen}, it is used
for permutation and boostsrap procedures.
}
\examples{
\dontrun{
if(interactive()){
 squared.ev <- sv2(matrix(runif(15), nrow = 3))
 }
}
}
\author{
Hervé Abdi
}
