% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tepCCA.R
\name{tepCCA}
\alias{tepCCA}
\title{Title}
\usage{
tepCCA(DATA1, DATA2, center1 = TRUE, scale1 = "SS1", center2 = TRUE,
  scale2 = "SS1", DESIGN = NULL, make_design_nominal = TRUE,
  graphs = TRUE, k = 0)
}
\arguments{
\item{DATA1}{}

\item{DATA2}{}

\item{center1}{}

\item{scale1}{}

\item{center2}{}

\item{scale2}{}

\item{DESIGN}{}

\item{make_design_nominal}{}

\item{graphs}{}

\item{k}{}
}
\description{
Title
}
\examples{
\dontrun{}
}
