% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ptca.R
\name{theme_ptca}
\alias{theme_ptca}
\title{A \code{ggplot2} theme that matches the defaults
of the factorial figures in the package \code{PTCA4CATA}}
\usage{
theme_ptca(mire = TRUE, x0 = 0, y0 = 0, coor.fixed = TRUE,
  base_family = "")
}
\arguments{
\item{mire}{(Default = \code{TRUE}) when \code{TRUE}
plot the factorial axis at values \code{x0} and \code{y0}.}

\item{x0}{(Default = 0) the X-axis origin. Relevant only
when \code{mire = TRUE}.}

\item{y0}{(Default = 0) the Y-axis origin.
Relevant only
when \code{mire = TRUE}.}

\item{coor.fixed}{(Default = \code{TRUE}), when \code{TRUE}
 the X and Y axis have the same units, when \code{FALSE} 
the units are set by \code{ggplot2}, if a number is provided
it will give the ratio (i.e., Y/X) to be used.}

\item{base_family}{the font family for the text 
(e.g., on a mac 'Times', 'Arial', 'Courier', 'Symbol' ,
or plain R, 'sans', 'serif', 'mono', can be other fonts 
if installed). Default is "" (i.e., uses the current default)}
}
\description{
\code{theme_ptca}:
A \code{ggplot2} theme that matches the defaults
of the factorial figures in \code{PTCA4CATA}.
}
\examples{
\dontrun{
x <- c(5, 8, 1, 70)
y <- c(10, 12, 20, 50)
(p <-  ggplot(as.data.frame(cbind(x, y)),
              aes(x = x, y = y)) +
    geom_point(size = 2, shape = 19) +  coord_fixed() +
    theme_ptca())
}
}
\author{
Vincent Guillemot & Hervé Abdi
}
