% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ptca.R
\name{theme_ptca2}
\alias{theme_ptca2}
\title{A \code{ggplot2} theme that matches the defaults
of the factorial figures in the package \code{PTCA4CATA}}
\usage{
theme_ptca2(mire = TRUE, x0 = 0, y0 = 0, coor.fixed = TRUE,
  base_family = "")
}
\arguments{
\item{mire}{(Default = \code{TRUE}) when \code{TRUE}
plot the factorial axis at values \code{x0} and \code{y0}.}

\item{x0}{(Default = 0) the X-axis origin. Relevant only
when \code{mire = TRUE}.}

\item{y0}{(Default = 0) the Y-axis origin.
Relevant only
when \code{mire = TRUE}.}

\item{coor.fixed}{(Default = \code{TRUE}), when \code{TRUE}
 the X and Y axis have the same units, when \code{FALSE} 
the units are set by \code{ggplot2}, if a number is provided
it will give the ratio (i.e., Y/X) to be used.}

\item{base_family}{the font family for the text 
(e.g., on a mac \code{'Times', 'Arial', 'Courier', 'Symbol'},
or plain \code{R, 'sans', 'serif', 'mono'}, can also be other fonts 
if installed). Default is "" (i.e., uses the current default).}
}
\description{
\code{theme_ptca2}:
A \code{ggplot2} theme that matches the defaults
of the factorial figures in \code{PTCA4CATA}.
\code{theme_ptca2} is obviously a (slightly smarter) 
cousin of 
\code{PTCA4CATA::theme_ptca} (which will be deprecated in the
new future).
}
\examples{
\dontrun{
x <- c(5, 8, 1, 70)
y <- c(10, 12, 20, 50)
( # parentheses are for print
 p <-  ggplot(as.data.frame(cbind(x, y)), aes(x = x, y = y)) +
           geom_point(size = 2, shape = 19) +
           theme_ptca2() 
           )
}
}
\author{
Vincent Guillemot & Hervé Abdi
}
