% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_internal_functions.R
\name{tolerance.svd}
\alias{tolerance.svd}
\title{tolerance.svd, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN}
\usage{
tolerance.svd(x, nu = min(dim(x)), nv = min(dim(x)),
  tol = .Machine$double.eps)
}
\description{
tolerance.svd, authored by Derek Beaton. Cf. package GSVD on GitHub, soon available on the CRAN
}
\keyword{internal}
