% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twentyWines.R
\docType{data}
\name{twentyWines}
\alias{twentyWines}
\title{20 red wines (10 French, 10 American) are rated by an expert
on multiple dimensions.}
\format{
A list containing 3 data frames:
\itemize{
\item \code{df.active} the ratings of the 20 wines for
\code{Sugar} and \code{Astringent}.
\item \code{supplementary.variables} stores the long names of the wines,
and the ratings of the 20 wines
for the \emph{supplementary} variables  \code{Acid, Bitter, Fruity, Woody},
and the \code{Origin} of the wines (French vs USA).
\item \code{supplementary.observation} the ratings for the
mystery wine (Trius Red: a wine from Ontario).
}
}
\usage{
data("twentyWines")
}
\description{
\code{twentyWines}:
20 (imaginary) red wines (10 French, 10 American) are rated,
from 1 to 20,
on multiple dimensions, by a (semi-fictitious) expert.
}
\details{
The data are derived from real data but have been
"massaged"
to be nice (e.g., mean, variances, eigenvalues, are integers).
The 10 French wines were a mix of
Carbernet Sauvignon and Merlot and the 10 American wines
were made from Zinfandel.

The wines were first evaluated on \code{Sugar} and
\code{Astringent}; these 2 dimensions are stored in the
data frame \code{df.active}.

After the evaluation of the 20 wines was performed,
the expert was also
asked to evaluate (on the same dimensions) another red wine
(\emph{Trius}: a Cabernet-Merlot from Ontario, Canada). The results
for this wine are stored in the data frame:
\code{supplementary.observation}

After the first evaluation of the 21 wines,
the (semi-fictitious) expert rated the
wine on 2 other dimensions: \code{Acid} and \code{Bitter}.

Incidently, the expert (even though semi-fictitious) spontaneously
commented that some wines were \code{Fruity} or \code{Woody}.

The long names of the wines, the origin of the wines
(\code{F} for French and \code{U} for USA), the additional ratings,
and the comments are stored in the data frame
\code{supplementary.variables}.
}
\references{
These ratings were (strongly) inspired by the
the \emph{Words} example from:

Abdi, H., Edelman, B., Valentin, D., & Dowling, W.J. (2009).
\emph{Experimental Design and Analysis for Psychology}.
Oxford: Oxford University Press.
}
\author{
Hervé Abdi and Dominique Valentin
}
\keyword{data4PCCAR}
\keyword{datasets}
