% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winesOf3Colors.R
\docType{data}
\name{winesOf3Colors}
\alias{winesOf3Colors}
\title{36 wines are evaluated on several dimensions orgainzed in 3 blocks.}
\format{
A list with
one data frame with three blocks of data:
\itemize{
\item \code{1} The descriptors.
\item \code{2} The chemistry
\item \code{3} The sensory
}
}
\usage{
data("winesOf3Colors")
}
\description{
\code{winesOf3Colors}:
36 wines are evaluated on several dimensions orgainzed in 3 blocks.
}
\details{
The (fictitious) wines come form 3 countries
Argentina, Canada and the USA. They also come in three different
colors: Red, Rosé, and White. They are decribed by
they characteristic, chemistry and sensory properties
}
\references{
Thesedata have been used to illustrate
PLSC and CCA
}
\author{
Hervé Abdi and Dominique Valentin
}
\keyword{data4PCCAR}
\keyword{datasets}
