% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiplot4PCA.R
\name{addArrowsAndNames}
\alias{addArrowsAndNames}
\title{Add arrows and names
to a factorial graph created by
\code{\link[PTCA4CATA]{createFactorMap}}
and friends.}
\usage{
addArrowsAndNames(
  DATA,
  axis1 = 1,
  axis2 = 2,
  fontface = "bold.italic",
  color = "darkolivegreen"
)
}
\arguments{
\item{DATA}{a matrix or data frame with coordinates
of the items to be plotted
(NB \emph{needs to have row names}).}

\item{axis1}{horizontal axis (\code{Default: 1}).}

\item{axis2}{vertical axis (\code{Default: 2}).}

\item{fontface}{font face for the names
(\code{Default: 'bold.italic'}).}

\item{color}{an element or
a vector of color (\code{Default: 'darkolivegreen'}).}
}
\value{
A list with two elements
\enumerate{
\item \code{arrows}  (\code{ggplot} code to add the arrows)
\item \code{labels4Biplot}
(\code{ggplot} code to add the names of the
items): Uses \code{ggplot2::annotate()}.
}
}
\description{
\code{addArrowsAndNames}:
Add arrows and names
to a factorial graph created by
\code{\link[PTCA4CATA]{createFactorMap}}
and friends. Use, for example, to create
a \emph{biplot} by adding variables
to the observation map
}
\details{
To be used as an add-on for map
created by  \code{\link[PTCA4CATA]{createFactorMap}}.
Note that for strange reasons (i.e., \code{ggplot2} internals ....),
the returned list cannot be provided with a \code{class}
(and so the \code{print} function cannot give a help)
}
\examples{
\dontrun{
data("twentyWines") # get the wine data set
resPCA <- ExPosition::epPCA(twentyWines$df.active, 
                            scale = FALSE, graphs = FALSE)
getCoordinates <- getBiplotCoor(resPCA) # get biplot coordinates
aMap <- PTCA4CATA::createFactorMap(resPCA$ExPosition.Data$fi, 
         constraints = getCoordinates$constraints4Biplot) # create base map
add2Map <- addArrowsAndNames(getCoordinates$coordArrow) # get the arrows
print(aMap$zeMap + add2Map) # print the biplot map
}
}
\seealso{
\code{\link{getBiplotCoor}}
}
\author{
Hervé Abdi
}
