% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4PCCAR.R
\docType{package}
\name{data4PCCAR}
\alias{data4PCCAR}
\alias{data4PCCAR-package}
\title{A collection of data sets and \eqn{R}-functions to illustrate
and complement:  Abdi, H., & Beaton, D. (2019)
\emph{Principal Component and Correspondence Analyses with R}.
New York: Springer.}
\description{
\code{data4PCCAR:}  These data sets and \eqn{R}-functions are used
to illustrate principal component analysis,
correspondence analysis, and multidimensional scaling.
\tabular{ll}{ Package: \tab datarPCCAR\cr Type: \tab Package\cr Version: \tab
0.1.0\cr Date: \tab 2018-03-15\cr License: \tab GPL-2\cr Depends: \tab
 (>= 3.0.0)}
}
\note{
The current version of data4PPCCAR needs to be installed from
`Github` with the command
`devtools::install_github('HerveAbdi/data4PCCAR')``
}
\author{
Herve Abdi [aut, com, ctb, cre],
Vincent Guillemot  [aut, cre],
& Derek Beaton
[aut, cre]

Maintainer: Herve Abdi <herve@utdallas.edu>
}
\keyword{package}
