% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpos.R
\name{firstpos}
\alias{firstpos}
\title{ensures that the first element of all the columns
of one/two matrix/matrices
is positive.}
\usage{
firstpos(P, Q = NULL)
}
\arguments{
\item{P}{A matrix of numeric values}

\item{Q}{If not \code{NULL}, 
a matrix of numeric values with the 
same number of columns as \code{P}}
}
\value{
If \code{Q} is \code{NULL}, 
then the function returns a only the modified P matrix.
Otherwise, the functions returns a list with two elements:
\itemize{
  \item \code{P} A matrix with the same values as in \code{P}, 
  except that the sign of all the values in each column
   is flipped if the first value of this column is negative.
  \item \code{Q} A matrix with the same values as in 
  \code{Q}, 
  except that the sign of all the values in each column 
  is flipped the same way that it is flipped for\code{P}.
}
}
\description{
\code{firspos}: 
Changes the sign of the values of the columns 
in one or two matrices 
according to the sign of the elements in the first line 
of the first matrix. 
After transformation the first element of every column of the
matric or matrices will be positive.

This is used to get reproducible 
results when performing singular value decompositions 
or eigenvalue decompositions.

Note that when provided tow matrices the signs of the
matrices will match: This is used to unsures that
singular vectors of a given matrix match.
}
\examples{
\dontrun{
 A <- matrix(rnorm(4), 2, 2)
 B <- matrix(rnorm(6), 3, 2)
 firstpos(A, B)
}
}
\seealso{
firstpos4ExPo
}
\author{
Hervé Abdi
}
