% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveBeers5Judges.R
\docType{data}
\name{fiveBeers5Judges}
\alias{fiveBeers5Judges}
\title{Five tasters evaluated
five beers on dimensions
that they  had previously chosen.
These data can be used to illustrate
Multiple Factor Analysis (MFA) or Distatis.}
\format{
A list containing 1 data frames:
\itemize{
\item \code{ratings:} A 5 by 5 data frame with
the original ratings performed using a 10 point rating scale.
\item \code{fuzzyRatings:}
A 5 by 26 data frame with
the beers being rows and the evalutions
by the judges begin the columns.
}
}
\usage{
data("fiveBeers5Judges")
}
\description{
\code{fiveBeers5Judges}:
Five tasters evaluated
five beers on dimensions
that they had previously chosen.
These data can be used to illustrate
Multiple Factor Analysis (MFA) or Distatis.

The judges are called \code{J1}
to \code{J5}
}
\references{
These data were used in a couple of workshops
for the SPISE meeting.
}
\author{
Hervé Abdi
}
\keyword{data4PCCAR}
\keyword{datasets}
