% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodOfTheWorld.R
\docType{data}
\name{foodOfTheWorld}
\alias{foodOfTheWorld}
\title{26 world cuisines are described by
their cooking ingredients.}
\format{A list containing 9 objects:
 1) \code{CT}: a 26 cuisines by 159 ingredients data frame
 containing the contingency table;
 2) \code{logCT}: a 26 cuisines by 159 ingredients
 data frame containing the "logged" contingency table;
 3) \code{logCT.integer}: a 26 cuisines by 159 ingredients
 data frame containing the rounded"logged" contingency table;
 4) \code{small.CT}: a 26 cuisines by 68 ingredients data frame
 containing the "small" contingency table;
 5) \code{small.logCT}: a 26 cuisines by 68 ingredients data frame
 containing the "small" "logged" contingency table;
 6) \code{small.logCT.integer}:
 a 26 cuisines by 68 ingredients data frame
 containing the "small" rounded "logged" contingency table;
 7) \code{compact.CT}: a 26 cuisines by 22 ingredients data frame
 containing the "compact" contingency table;
 8) \code{compact.logCT}: a 26 cuisines by 22 ingredients data frame
 containing the "compact" "logged" contingency table; and
 9) \code{compact.logCT.integer}:
 a 26 cuisines by 22 ingredients data frame
 containing the "compact" rounded "logged" contingency table;
 @keywords datasets data4PCCAR
 @author Herve Abdi & Jyothi}
\usage{
data("foodOfTheWorld")
}
\description{
\code{foodOfTheWorld}: 26 world cuisines are described by
their cooking ingredients. The long form of the data comprises
159 ingredients. A shorter form includes the 68 most relevant
ingredients.
}
\details{
These data sets are derived from the work of blogger
\emph{Jyothi}
(see \code{https://www.r-bloggers.com/a-visualization-of-world-cuisines/}).
Here the main data
consist in a list with three main data frames:
 1) \code{CT}: a contingency table listing
the number of recipes (from the publicly available
data base \code{Epicurious}) that mention a given ingredient
for a given cuisine;
2) \code{small.CT}
A short version is also provided that is
using only 68
(most discriminant) ingredients;
and 3) \code{small.CT}
a compact version
(that was given in the original post) with only 22
(most informative) ingredients.
Because the counts vary greatly,
the contingency tables are also re-scaled by taking the
logarithm of the contingency table [with log(0) set to 0].
The log tables are also available as a pseudo-integer version
(obtained by multiplying by 10 and rounding the "logged"
contingency table). The log rescaled tables are preceded
by the prefix \code{log}, the "log-rounded" tables
are, in addition,  followed by the suffix \code{.interger}

The original post used these data  (the compact version)
to illustrate
principal component analysis, but because these data are count
\emph{correspondence analysis} is more appropriate.

The large, small, and compact data sets give roughly the
same first dimension(s), whereas versions
with more deescriptors give
more signfificant/relevant dimensions than data sets with fewer
variables.
}
