% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomImage4PCA.R
\name{getRandomFS}
\alias{getRandomFS}
\title{generate multivariate random observations (i.e.,
factor scores) whose variance
 match the variance/eigenvalues of the dimensions of the space.}
\usage{
getRandomFS(ev, nObs = 100, FUN = stats::runif, center = FALSE,
  scale = "SS1", sv = FALSE)
}
\arguments{
\item{ev}{The variance / eigenvalues per dimension (no default).}

\item{nObs}{the number of observationz to generate, Default: 100}

\item{FUN}{the random number generating.
function. Default: \code{stats::runif} (uniform distribution).}

\item{center}{center the numbers.  Default: \code{FALSE}.}

\item{scale}{normalization per dimension prior to
re-normalize the factor scores with \code{ev}, best to
stick with the default.
if \code{FALSE} do not normalize
the numbers, if \code{TRUE} normalize the numbers to Z-scores
per dimension, if \code{'SS1'} (default) normalize to one}

\item{sv}{if \code{TRUE}, \code{ev} gives singular values or
standard deviation, if
  \code{FALSE} (default)
  \code{ev} gives eigenvalues or
variance.}
}
\value{
A matrix
of multidimensional factor scores
with \code{nObs} rows (observations) and number of columns
(variables / components)
equal to
 the length of \code{ev}.
}
\description{
\code{getRandomFS}:
generate random observations
(such as factors scores) from a PCA whose variance
match the variance of the dimensions of the PCA.
.
}
\details{
\code{getRandomFS}: Simulates a set
multivariate coordinates of random observations
whose variance per dimension
(i.e.,
eigenvalues in PCA) will match a given set of variances.
The function generating the set of numbers 
can be any such function
(default is \code{runif}).
}
\examples{
# generates 10 factors scores from a 4-dimensional space
# whose dimensions have variances of 16, 9, 4, and 1
randFS <- getRandomFS(ev = c(16, 9, 4, 1), nObs = 10)
}
\seealso{
\code{\link[data4PCCAR]{scale0}}
 @importFrom stats runif
}
\author{
Hervé Abdi
}
