% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS_jack_svds_HA.R
\name{normaliz}
\alias{normaliz}
\title{normalize and send back a matrix normalized by column
(i.e., each column vector has a norm of 1).}
\usage{
normaliz(F)
}
\arguments{
\item{F}{a matrix}
}
\value{
the column normalized version of \code{F}
}
\description{
\code{normaliz: }
normalizes and send back a matrix normalized by column
(i.e., each column vector has a norm of 1).
\code{normaliz} is used by \code{\link{PLSR_SVD}}.
}
\examples{
\dontrun{
X <-  matrix(c(4,2,8,7), nrow = 2)
Z <- normaliz(X)
}
}
\author{
Hervé Abdi, Lei Xuan
}
