% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplementary4Vari.R
\name{print.epVariSup.i}
\alias{print.epVariSup.i}
\title{Change the print function for createFactorMap}
\usage{
\method{print}{epVariSup.i}(x, ...)
}
\arguments{
\item{x}{a list: output of createFactorMap}

\item{...}{everything else for the functions}
}
\description{
Change the print function for createFactorMap
}
\author{
Hervé Abdi
}
