% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS_jack_svds_HA.R
\name{print.PLSR_SVD}
\alias{print.PLSR_SVD}
\title{print function for \code{print.PLSR_SVD}}
\usage{
\method{print}{PLSR_SVD}(x, ...)
}
\arguments{
\item{x}{The input}

\item{...}{everything else}
}
\value{
printed help
}
\description{
print function for \code{print.PLSR_SVD} objects.
}
\seealso{
\code{\link{PLSR_SVD}}
}
\author{
Hervé Abdi
}
