% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorsOfVowels.R
\docType{data}
\name{sixFaces}
\alias{sixFaces}
\title{Six two-dimensional images of faces, three males and three females.}
\format{An array
#'  @keywords datasets data4PCCAR
 @author Herve Abdi
 @references
 Add reference}
\usage{
data("sixFaces")
}
\description{
\code{sixFaces}
Six two dimensional images of faces.
}
\details{
	Six two dimensional images of faces.}
