% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary4PLSC.R
\name{supplementaryObservations4PLSC}
\alias{supplementaryObservations4PLSC}
\title{compute the latent variables for
supplementary observations for a PLSC model
computed with  \code{\link[TExPosition]{tepPLS}}.}
\usage{
supplementaryObservations4PLSC(resPLSC, Xsup = NULL, Ysup = NULL)
}
\arguments{
\item{resPLSC}{the results of a \code{PLSC} analysis
from  \code{\link[TExPosition]{tepPLS}}.}

\item{Xsup}{an \eqn{N}sup by \eqn{I} matrix of
supplementary observations matching the \strong{X} matrix
(see \code{description for details}).
When \code{NULL} (Default) nothing is computed for \code{Xsup}.}

\item{Ysup}{an \eqn{N}sup by \eqn{J} matrix of
supplementary observations matching the \strong{Y} matrix
(see \code{description for details}).
When \code{NULL} (Default) nothing is computed for \code{Ysup}.}
}
\value{
A list with \code{lx.sup} and \code{ly.sup}
giving the latent variables values
of the supplementary observations
for (respectively) \eqn{X} and \eqn{Y}.
}
\description{
\code{supplementaryObservations4PLSC}:
computes the latent variables for
supplementary observations for a PLSC model
computed with  \code{\link[TExPosition]{tepPLS}}.
}
\details{
The original analysis is performed with
\code{\link[TExPosition]{tepPLS}} on the original data matrices
\strong{X} (\eqn{N} by \eqn{I}) and \strong{Y} (\eqn{N} by \eqn{J}).
The supplementary data matrices should have \eqn{I}
columns for \eqn{X}sup and \eqn{J}
columns for \eqn{Y}sup.
}
