% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementaryObservations4PLSCA.R
\name{supplementaryObservations4PLSCA}
\alias{supplementaryObservations4PLSCA}
\title{compute the latent variables for
supplementary observations for a PLSC model
computed with  \code{\link[TExPosition]{tepPLSCA}}.}
\usage{
supplementaryObservations4PLSCA(resPLSCA, Xsup = NULL, Ysup = NULL)
}
\arguments{
\item{resPLSCA}{the results of a \code{PLSCA} analysis
from  \code{\link[TExPosition]{tepPLSCA}}.}

\item{Xsup}{an \eqn{N}sup by \eqn{I} matrix of
supplementary observations matching the \strong{X} matrix
(see \code{description} for details).
When \code{NULL} (Default) nothing is computed for \code{Xsup}.}

\item{Ysup}{an \eqn{N}sup by \eqn{J} matrix of
supplementary observations matching the \strong{Y} matrix
(see \code{description} for details).
When \code{NULL} (Default) nothing is computed for \code{Ysup}.}
}
\value{
A list with \code{lx.sup} and \code{ly.sup}
giving the latent variables values
of the supplementary observations
for (respectively) \eqn{X} and \eqn{Y}.
}
\description{
\code{supplementaryObservations4PLSCA}:
computes the latent variables for
supplementary observations for a PLSCA model
computed with  \code{\link[TExPosition]{tepPLSCA}}.
}
\details{
The original analysis is performed with
\code{\link[TExPosition]{tepPLSCA}} on the original data matrices
\strong{X} (\eqn{N} by \eqn{I}) and \strong{Y} (\eqn{N} by \eqn{J}).
The supplementary data matrices should have \eqn{I}
columns for \eqn{X}sup and \eqn{J}
columns for \eqn{Y}sup. Note that \code{PLSCA} is used
with qualitative variables (i.e., factors) recoded
as 0/1 variables with disjunctive coding
(i.e., with  \code{\link[ExPosition]{makeNominalData}}),
the supplementary
variables need to be recoded in the same way.
\subsection{Implementation}{

For \code{PLSCA} the observations needs to be pre-processed in
the same way as the original observations. Often in \code{PLSCA}
the observations are described by qualitative variables
(in general coded as \emph{factors}) by \code{R}.
When this is the case the supplementary observations should coded
as factors too with the same levels (aka modalities) as
the active observations
(see also  \code{\link[ExPosition]{makeNominalData}}).
}

\subsection{Computation}{

The projections of supplementary observations in \code{PLSC}
is obtained using the standard \emph{transition formulas}
from correspondence analysis.
}
}
\seealso{
\code{\link[TExPosition]{tepPLSCA}}
\code{\link[ExPosition]{makeRowProfiles}}
\code{\link{supplementaryVariables4PLSCA}}
\code{\link{supplementaryObservations4PLSC}}
}
\author{
Hervé Abdi
#' @references
See:

Beaton, D., Dunlop, J., ADNI, & Abdi, H. (2016).
Partial Least Squares-Correspondence Analysis (PLSCA):
A framework to simultaneously analyze behavioral and genetic data.
\emph{Psychological Methods, 21}, 621-651.

Abdi H. & Béra, M. (2018).
Correspondence analysis.
In R. Alhajj and J. Rokne (Eds.),
\emph{Encyclopedia of Social Networks and Mining (2nd Edition)}.
New York: Springer Verlag.

Abdi, H. (2007). Singular Value Decomposition (SVD)
and Generalized Singular Value Decomposition (GSVD).
In N.J. Salkind (Ed.):
\emph{Encyclopedia of Measurement and Statistics}.
Thousand Oaks (CA): Sage. pp. 907-912.
}
